Abstract Actions

General concept

Parameters (possibly region- or item-dependent) are handle by AbstractAction types and can be assigned to Assembly Patterns. The action is applied to some input vector, that collects the evaluations of the function operators of (a well-defined subset of the) arguments of the assembly pattern, and then returns the modified result. Each AbstractAction implements the apply! interface

apply!(result,input,::AbstractAction,qp::Int)

that defines how the result vector is computed based on the input and the current quadrature point number.

Before it is applied to any input or quadrature point, an update! step is performed upon entry on any item which prepares the data for the evaluation (fixing item and region numbers or computing the global coordinates of the quadrature points). Details of that are hidden here, since it is usually not necessary for the user to interfere with that.

Below is a table with implemented action types and some short description what it can be used for. Click on the action name for more details.

Action type	What it does

DoNotChangeAction	does nothing (result = input)

MultiplyScalarAction	multiplies argument (vector) with a scalar

RegionWiseMultiplyScalarAction	multiplies argument (vector) with a scalar that depends on the region

MultiplyVectorAction	multiplies argument (vector) componentwise with a vector of same length

RegionWiseMultiplyVectorAction	multiplies argument (vector) componentwise with a region-depending vector of same length

MultiplyMatrixAction	matrix-vector multiplication, input is multiplied with the specified matrix

FunctionAction	f!(result, input) for some user-specified function with this interface

XFunctionAction	f!(result, input, x) for some user-specified function (that also depends on x)

ItemWiseXFunctionAction	f!(result, input, x, item) for some user-specified function (that also depends on x and item number)

RegionWiseXFunctionAction	f!(result, input, x, region) for some user-specified function (that also depends on x and region id)

GradientRobustMultiPhysics.DoNotChangeAction — Type

struct DoNotChangeAction <: AbstractAction

action that does nothing to the input, i.e.

result[j] = input[j]

(for j = 1 : argsizes[1])

GradientRobustMultiPhysics.DoNotChangeAction — Method

function DoNotChangeAction(ncomponents::Int)

creates DoNotChangeAction with specified argsizes = (ncomponents, ncomponents).

GradientRobustMultiPhysics.FunctionAction — Type

struct FunctionAction{T<:Real} <: AbstractAction

action that puts input into the specified function f! and returns its result. The function f! should have the interface

f!(result,input)

and result is expected to be of length argsizes[1], input of length argsizes[2].

The quadrature order of assemblies that involve this action can be altered with bonus_quadorder.

GradientRobustMultiPhysics.FunctionAction — Method

function FunctionAction(f!::Function, argsizes::Array{Int,1}; bonus_quadorder::Int = 0)

creates FunctionAction.

GradientRobustMultiPhysics.ItemWiseFunctionAction — Type

mutable struct ItemWiseFunctionAction{T<:Real} <: AbstractAction

action that puts input into the specified function f! and returns its result. The function f! should have the interface

f!(result,input,item)

and result is expected to be of length argsizes[1], input of length argsizes[2].

The quadrature order of assemblies that involve this action can be altered with bonus_quadorder.

GradientRobustMultiPhysics.ItemWiseFunctionAction — Method

function ItemWiseFunctionAction(f!::Function, argsizes::Array{Int,1}; bonus_quadorder::Int = 0)

creates ItemWiseFunctionAction.

GradientRobustMultiPhysics.ItemWiseMultiplyScalarAction — Type

mutable struct ItemWiseMultiplyScalarAction{T<:Real} <: AbstractAction

action that multiplies an item-dependent scalar to the input, i.e.

result[j] = input[j] * value[item]

(for j = 1 : argsizes[1])

GradientRobustMultiPhysics.ItemWiseMultiplyScalarAction — Method

function ItemWiseMultiplyScalarAction(value4item::Array{<:Real,1}, argsizes::Array{Int,1})

creates ItemWiseMultiplyScalarAction with specified argsizes = (length(result), length(input)).

GradientRobustMultiPhysics.ItemWiseXFunctionAction — Type

mutable struct ItemWiseXFunctionAction{T<:Real} <: AbstractAction

action that puts input into the specified function f! and returns its result. The function f! should have the interface

f!(result,input,x,item)

and result is expected to be of length argsizes[1], input of length argsizes[2].

The quadrature order of assemblies that involve this action can be altered with bonus_quadorder.

GradientRobustMultiPhysics.ItemWiseXFunctionAction — Method

function ItemWiseXFunctionAction(f!::Function, argsizes::Array{Int,1}, xdim::Int; bonus_quadorder::Int = 0)

creates ItemWiseXFunctionAction.

GradientRobustMultiPhysics.MultiplyMatrixAction — Type

struct MultiplyMatrixAction{T<:Real} <: AbstractAction

action that multiplies input vector with a matrix, i.e. (note that also gradients are transferred as vectors)

result = value * input

(resultdim = argsizes[1])

GradientRobustMultiPhysics.MultiplyMatrixAction — Method

function MultiplyMatrixAction(matrix::Array{<:Real,2}, argsizes::Array{Int,1})

creates MultiplyMatrixAction with specified argsizes = (length(result), length(input)).

GradientRobustMultiPhysics.MultiplyScalarAction — Type

struct MultiplyScalarAction{T<:Real} <: AbstractAction

action that multiplies a scalar to the input, i.e.

result[j] = input[j] * value

(for j = 1 : argsizes[1])

GradientRobustMultiPhysics.MultiplyScalarAction — Method

function MultiplyScalarAction(value::Real, argsizes::Array{Int,1})

creates MultiplyScalarAction with specified argsizes = (length(result), length(input)).

GradientRobustMultiPhysics.MultiplyVectorAction — Type

struct MultiplyVectorAction{T<:Real} <: AbstractAction

action that multiplies a different scalar to each component of the input, i.e.

result[j] = input[j] * value[j]

(for j = 1 : argsizes[1])

GradientRobustMultiPhysics.MultiplyVectorAction — Method

function MultiplyVectorAction(values::Array{<:Real,1}, argsizes::Array{Int,1})

creates MultiplyVectorAction with specified argsizes = (length(result), length(input)).

GradientRobustMultiPhysics.RegionWiseMultiplyScalarAction — Type

struct RegionWiseMultiplyScalarAction{T<:Real} <: AbstractAction

action that multiplies a region-dependent scalar to the input, i.e.

result[j] = input[j] * value[region]

(for j = 1 : argsizes[1])

GradientRobustMultiPhysics.RegionWiseMultiplyScalarAction — Method

function RegionWiseMultiplyScalarAction(value4region::Array{<:Real,1}, argsizes::Array{Int,1})

creates RegionWiseMultiplyScalarAction with specified argsizes = (length(result), length(input)).

GradientRobustMultiPhysics.RegionWiseMultiplyVectorAction — Type

struct RegionWiseMultiplyVectorAction{T<:Real} <: AbstractAction

action that multiplies a different region-dependent scalar to each component of the input, i.e.

result[j] = input[j] * value[region][j]

(for j = 1 : argsizes[1])

GradientRobustMultiPhysics.RegionWiseMultiplyVectorAction — Method

function RegionWiseMultiplyVectorAction(values4region::Array{Array{<:Real,1},1}, argsizes::Array{Int,1})

creates RegionWiseMultiplyVectorAction with specified argsizes = (length(result), length(input)).

GradientRobustMultiPhysics.RegionWiseXFunctionAction — Type

mutable struct RegionWiseXFunctionAction{T<:Real} <: AbstractAction

action that puts input into the specified function f! and returns its result. The function f! should have the interface

f!(result,input,x,region)

and result is expected to be of length argsizes[1], input of length argsizes[2].

The quadrature order of assemblies that involve this action can be altered with bonus_quadorder.

GradientRobustMultiPhysics.RegionWiseXFunctionAction — Method

function RegionWiseXFunctionAction(f!::Function, argsizes::Array{Int,1}, xdim::Int; bonus_quadorder::Int = 0)

creates RegionWiseXFunctionAction.

GradientRobustMultiPhysics.XFunctionAction — Type

struct XFunctionAction{T<:Real} <: AbstractAction

action that puts input into the specified function f! and returns its result. The function f! should have the interface

f!(result,input,x)

and result is expected to be of length argsizes[1], input of length argsizes[2].

The quadrature order of assemblies that involve this action can be altered with bonus_quadorder.

GradientRobustMultiPhysics.XFunctionAction — Method

function XFunctionAction(f!::Function, argsizes::Array{Int,1}, xdim::Int; bonus_quadorder::Int = 0)

creates XFunctionAction.