Assembly Details

The assembly of an operator is essentially based on a combination of Assembly Types and Assembly Patterns. The assembly type mainly allows to choose the geometry information needed for providing quadrature and dof handling. The assembly pattern then basically evaluates the function operators and action for the ansatz and test functions and does the quadrature-weighted accumulation into matrices or vectors that represent the operators.

Assembly Types

The following assembly types are available. Additional to define where AssemblyPatterns live and assemble, they can be also used as an argument for interpolation!.

AssemblyType	Description

AT_NODES	interpolate at vertices of the mesh (only for H1-conforming FEM)

ON_CELLS	assemble/interpolate over the cells of the mesh

ON_FACES	assemble/interpolate over all faces of the mesh

ON_IFACES	assemble/interpolate over the interior faces of the mesh

ON_BFACES	assemble/interpolate over the boundary faces of the mesh

ON_EDGES (*)	assemble/interpolate over all edges of the mesh (in 3D)

ON_BEDGES (*)	assemble/interpolate over the boundary edges of the mesh (in 3D)

Note

(*) = only reasonable in 3D and still experimental, might have some issues

GradientRobustMultiPhysics.AT_NODES — Type

abstract type AT_NODES <: AbstractAssemblyType

causes interpolation at vertices of the grid (only for H1-conforming interpolations)

GradientRobustMultiPhysics.ON_BEDGES — Type

abstract type ON_BEDGES <: AbstractAssemblyType

causes assembly/interpolation on boundary edges of the grid (only in 3D)

GradientRobustMultiPhysics.ON_BFACES — Type

abstract type ON_BFACES <: AbstractAssemblyType

causes assembly/interpolation on boundary faces of the grid

GradientRobustMultiPhysics.ON_CELLS — Type

abstract type ON_CELLS <: AbstractAssemblyType

causes assembly/interpolation on cells of the grid

GradientRobustMultiPhysics.ON_EDGES — Type

abstract type ON_EDGES <: AbstractAssemblyType

causes assembly/interpolation on edges of the grid (only in 3D)

GradientRobustMultiPhysics.ON_FACES — Type

abstract type ON_FACES <: AbstractAssemblyType

causes assembly/interpolation on faces of the grid

GradientRobustMultiPhysics.ON_IFACES — Type

abstract type ON_IFACES <: on faces

causes assembly/interpolation on interior faces of the grid

Assembly Patterns

Each Pattern comes with a number of arguments/quantities with associated Function Operators as well as one of the Assembly Types that states whether the form is evaluated over CELLS, FACES order BFACES (see above). Important note: this assembly type is relative to the grid of the first argument of the pattern. If this argument already lives ONFACES and the pattern is also ONFACES, it will ultimatively assemble on the faces of the faces (that are the edges of the grid with these faces). Moreover, patterns can have Abstract Actions that allow to make the evaluations parameter-, region- and/or function-dependent. Each pattern then has usually on to three implementation that writes into FEMatrix or FEVector (where e.g. a subset of arguments is fixed) or evaluates the pattern in the given FEVectorBlocks.

The patterns are used to assembly the PDE operators defined in a PDE Description. However, it is also possible for the user to use them directly, see e.g. the example Commuting Interpolators (2D).

GradientRobustMultiPhysics.AssemblyPattern — Type

mutable struct AssemblyPattern{APT<:AssemblyPatternType, T<:Real, AT<:AbstractAssemblyType}

each assembly pattern has one of the assembly pattern types (APT) that trigger different assemblies for the involved finite element spaces, operators and an assigned action. The assembly type (AT) determines if the assembly takes place on cells, faces or edges etc. (relatively to the assembly type of the first argument of the pattern)

The following table lists all available assembly patterns, their constuctor names and how they can be used for assembly or evaluations.

AssemblyPatternType	constructor	evaluate	assembly into matrix	assembly into vector

APT_ItemIntegrator	ItemIntegrator	yes	no	no

APT_LinearForm	LinearForm	no	no	yes

APT_BilinearForm	BilinearForm	no	yes	yes (1)

APT_TrilinearForm	TrilinearForm	no	yes (1)	yes (2)

APT_MultiLinearForm	MultilinearForm	no	no	yes (N-1)

APT_NonlinearForm	NonlinearForm	no	yes (L)	yes (L)

Number in brackets denotes the number of fixed arguments needed for this assembly, (L) means that a current solution is needed to evaluate (to evaluate the linearisation of the nonlinear form in this state). Evaluations of the other AssemblyPatterns may be possible in a future update, but currently have to be performed by maintaining a duplicate of the pattern rewritten as an ItemIntegrator.

Constructor details

Below all constructors are detailed.

GradientRobustMultiPhysics.ItemIntegrator — Function

function ItemIntegrator(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates an ItemIntegrator assembly pattern with the given operators and action etc.

GradientRobustMultiPhysics.LinearForm — Function

function LinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a LinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.BilinearForm — Function

function BilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a general BilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.TrilinearForm — Function

function TrilinearForm(
    T::Type{<:Real},
    FES::Array{FESpace,1},          
    operators::Array{DataType,1},
    action::AbstractAction;
    regions::Array{Int,1} = [0])

Creates a TrilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.MultilinearForm — Function

function MultilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a MultilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.NonlinearForm — Function

function NonlinearForm(
    T::Type{<:Real},
    FES::Array{FESpace,1},          # finite element spaces for each operator of the ansatz function and the last one refers to the test function
    operators::Array{DataType,1},   # operators that should be evaluated for the ansatz function and the last one refers to the test function
    action::AbstractAction;         # action that shoul have an AbstractNLActionKernel
    regions::Array{Int,1} = [0])

Creates a NonlinearForm assembly pattern.

Evaluate and Assemble

Below all evaluate! and assemble! functions of the patterns are listed.

GradientRobustMultiPhysics.evaluate! — Function

function evaluate!(
    b::AbstractArray{T,2},          # target vector
    AP::AssemblyPattern{APT,T,AT},  # ItemIntegrator pattern
    FEB::Array{<:FEVectorBlock,1}   # coefficients for arguments
    where {APT <: APT_ItemIntegrator, T, AT}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlocks FEB into given two-dimensional Array b.

GradientRobustMultiPhysics.evaluate — Function

function evaluate(
    AP::AssemblyPattern{APT,T,AT},  # ItemIntegrator pattern
    FEB::Array{<:FEVectorBlock,1})  # coefficients for arguments
    where {APT <: APT_ItemIntegrator, T, AT}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlocks FEB, only returns accumulation over all items.

GradientRobustMultiPhysics.assemble! — Function

assemble!(
    b::Union{AbstractArray{T,1},AbstractArray{T,2}},    # target vector/matrix
    AP::AssemblyPattern{APT,T,AT};                      # LinearForm pattern
    factor = 1)                                         # factor that is multiplied
    where {APT <: APT_LinearForm, T, AT}

Assembly of a LinearForm pattern AP into a vector or matrix (if action is vetor-valued).

assemble!(
    A::AbstractArray{T,2},                  # target matrix
    AP::AssemblyPattern{APT,T,AT};          # BilinearForm Pattern
    apply_action_to::Int = 1,               # action is applied to which argument?
    factor = 1,                             # factor that is multiplied
    transposed_assembly::Bool = false,      # transpose result?
    transpose_copy = Nothing)               # copy a transposed block to this matrix
    where {APT <: APT_BilinearForm, T, AT}

Assembly of a BilinearForm BLF into given two-dimensional AbstractArray (e.g. FEMatrixBlock or a ExtendableSparseMatrix).

assemble!(
    b::AbstractArray{T,1},          # target vector
    AP::AssemblyPattern{APT,T,AT},  # BilinearForm Pattern
    fixedFE::AbstractArray;         # coefficients for fixed argument
    apply_action_to::Int = 1,       # action is applied to 1st or 2nd argument?
    fixed_arguments = [1],        # which argument is fixed?
    factor = 1)                     # factor that is multiplied
    where {APT <: APT_BilinearForm, T, AT}

Assembly of a BilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the second argument is fixed (default) by the given coefficients in fixedFE. With applyactionto=2 the action can be also applied to the second argument instead of the first one (default).

assemble!(
    assemble!(
    A::AbstractArray{T,2},                  # target matrix
    AP::AssemblyPattern{APT,T,AT},          # TrilinearForm pattern
    fixedFE::Array{<:FEVectorBlock,1};      # coefficients for fixed argument
    fixed_argument = [1],                   # position of fixed argument (as an array)
    transposed_assembly::Bool = false,      # transpose result?
    factor = 1)                             # factor that is multiplied
    where {APT <: APT_TrilinearForm, T, AT}

Assembly of a TrilinearForm AP into given two-dimensional AbstractArray (e.g. a FEMatrixBlock). Here, one argument (specified by fixed_argument) is fixed by the given coefficients in fixedFE[1]. Note, that the action is (currently) always applied to the first and second argument.

assemble!(
    assemble!(
    b::AbstractVector,                      # target vector
    AP::AssemblyPattern{APT,T,AT},          # coefficients for fixed arguments
    fixedFE::Array{<:FEVectorBlock,1};      # TrilinearForm pattern
    factor = 1)                             # factor that is multiplied
    where {APT <: APT_TrilinearForm, T, AT}

Assembly of a TrilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the first two arguments are fixed by the given coefficients in FE1 and FE2.

assemble!(
    b::AbstractVector,                        # target vector
    FE::Array{<:FEVectorBlock,1},             # coefficients of all but last argument
    AP::AssemblyPattern{APT,T,AT};            # Multilinearform pattern
    factor = 1)                               # factor that is multiplied
    where {APT <: APT_MultilinearForm, T, AT}

Assembly of a MultilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the all but the last arguments are fixed by the given coefficients in the components of FE.

assemble!(
    A::AbstractArray{T,2},                 # target matrix
    AP::AssemblyPattern{APT,T,AT};         # NonlinearForm pattern
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    factor = 1,                            # factor that is multiplied
    transposed_assembly::Bool = false)     # transpose result?
    where {APT <: APT_NonlinearForm, T, AT}

Assembly of a NonlinearForm assembly pattern into given two-dimensional AbstractArray (e.g. FEMatrixBlock).

assemble!(
    b::AbstractVector,                     # target vector
    AP::AssemblyPattern{APT,T,AT},         # NonlinearForm pattern
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    factor = 1)                            # factor that is multiplied
    where {APT <: APT_NonlinearForm, T, AT}

Assembly of a NonlinearForm AP into given AbstractVector (e.g. FEMatrixBlock).