Boundary Data

Dirichlet Boundary Data

BoundaryOperators carry the boundary data for each unknown. Each regions can have a different AbstractBoundaryType and an associated data function that satisfies the interface function data!(result,x::Array{<:Real,1}) or function data!(result,x::Array{<:Real,1},t::Real) if it is also time-dependent.

AbstractBoundaryType	Subtypes	causes

DirichletBoundary

BestapproxDirichletBoundary	computation of Dirichlet data by bestapproximation along boundary faces

InterpolateDirichletBoundary	computation of Dirichlet data by interpolation along boundary faces

HomogeneousDirichletBoundary	zero Dirichlet data on all dofs

GradientRobustMultiPhysics.BoundaryOperator — Type

struct BoundaryOperator <: AbstractPDEOperator

collects boundary data for a component of the system and allows to specify a AbstractBoundaryType for each boundary region so far only DirichletBoundary types (see above)

GradientRobustMultiPhysics.add_boundarydata! — Function

add_boundarydata!(PDE::PDEDescription, position::Int64, regions::Any, btype::Type{var"#s1564"} where var"#s1564"<:AbstractBoundaryType; data)

Adds the given boundary data with the specified AbstractBoundaryType at the specified position in the BoundaryOperator of the PDEDescription.

If timedependent == true, that data function depends also on time t and is reassembled in any advance! step of a TimeControlSolver.

Other Boundary Data

NeumannBoundary can be implemented as a RhsOperator with AT = ON_BFACES

PeriodicBoundary can be implemented as a CombineDofs <: AbstractGlobalConstraint

SymmetryBoundary can be implemented by penalisation as a AbstractBilinearForm on BFaces and specified boundary regions with operator NormalFlux + MultiplyScalarAction(penalty).