Data Export

Data export is possible into VTK (based on WriteVTK.jl) or CSV files.

GradientRobustMultiPhysics.writeCSV! — Method

writeCSV!(filename::String, Data::Vector{var"#s204"} where var"#s204"<:FEVectorBlock; operators, names, seperator)

Writes the specified FEVectorBlocks into a CSV datafile with the given filename. First d colomuns are the grid coordinates, the remaining columns are filled with the evaluations of the operators where operator[j] is applied to Data[j].

GradientRobustMultiPhysics.writeVTK! — Method

writeVTK!(filename::String, Data::Vector{var"#s206"} where var"#s206"<:FEVectorBlock; operators, names, vectorabs, add_regions, caplength)

Writes the specified FEVector into a vtk datafile with the given filename. Each FEVectorBlock in the Data array is saved as separate VTKPointData. Vector-valued quantities also generate a data field that represents the absolute value of the vector field at each grid point (if vectorabs is true).