FE Basis Evaluators

GradientRobustMultiPhysics.FEBasisEvaluator — Method

    FEBasisEvaluator{T <: Real,FEType <: AbstractFiniteElement,EG <: AbstractElementGeometry,FEOP <: AbstractFunctionOperator,AT <: AbstractAssemblyType}(FES::FESpace, qf::QuadratureRule)

Constructor for an evaluator for the basis of the specified FEType (with matching FESpace FES) of the function operator FEOP on the given element geometry EG beeing of AssemblyType AT at the points of the quadrature rule qf.

GradientRobustMultiPhysics.eval! — Method

    eval!(result, FEBE::FEBasisEvaluator, j::Int, i::Int, offset::Int = 0, factor = 1)

Evaluate the j-th basis function of the FEBasisEvaluator at the i-th quadrature point adn writes the (possibly vector-valued) evaluation into result (beginning at offset and with the specified factor).

GradientRobustMultiPhysics.eval! — Method

    eval!(result, FEBE::FEBasisEvaluator, j::Int, i::Int, offset::Int = 0, factor = 1)

Evaluates the linear combination of the basisfunction with given coefficients at the i-th quadrature point and writes the (possibly vector-valued) evaluation into result (beginning at offset and with the specified factor).

GradientRobustMultiPhysics.update! — Method

    update!(FEBE::FEBasisEvaluator, item::Int)

Update the FEBasisEvaluator on the given item number of the grid items associated to the AssemblyType. During the update the FEBasisevaluator computes all evaluations of all basis functions at all quadrature points and stores them in FEBE.cvals. From there they can be accessed directly or via the eval! functions.