Finite Element Spaces and Arrays

This page describes the structure FESpace that acts as a finite element space on a given grid. See Implemented Finite Elements for a list of available finite element types.

Moreover, there are special arrays FEVector and FEMatrix that carry coefficients and discretised PDEOperators.

FESpace

GradientRobustMultiPhysics.FESpace — Type

mutable struct FESpace{FEType<:AbstractFiniteElement}

A struct that has a finite element type as parameter and carries dofmaps (CellDofs, FaceDofs, BFaceDofs) plus additional grid information and access to arrays holding coefficients if needed.

Base.eltype — Method

eltype(_::FESpace{FEType<:AbstractFiniteElement}) -> Type{FEType} where FEType<:AbstractFiniteElement

Custom eltype function for FESpace returns the finite element type of the finite element space.

Base.show — Method

show(io::IO, FES::FESpace{FEType<:AbstractFiniteElement})

Custom show function for FESpace that prints some information and all available dofmaps.

FEVector

A FEVector consists of FEVectorBlocks that share a common one-dimensional arrays. Each block is associated to a FESpace and can only write into a region of the common array specified by offsets. It also acts as a one-dimensional AbstractArray itself.

GradientRobustMultiPhysics.FEVector — Type

struct FEVector{T} <: AbstractArray{T,1}

a plain array but with an additional layer of several FEVectorBlock subdivisions each carrying coefficients for their associated FESpace

GradientRobustMultiPhysics.FEVector — Method

FEVector{T}(name::String, FES::Array{FESpace,1}) where T <: Real

Creates FEVector that has one block for each FESpace in FES.

GradientRobustMultiPhysics.FEVector — Method

FEVector{T}(name::String, FES::FESpace) where T <: Real

Creates FEVector that has one block.

GradientRobustMultiPhysics.FEVectorBlock — Type

struct FEVectorBlock{T} <: AbstractArray{T,1}

block of an FEVector that carries coefficients for an associated FESpace and can be assigned as an AbstractArray (getindex, setindex, size, length)

Base.append! — Method

append!(FEF::FEVector{T<:Real}, name::String, FES::FESpace)

Custom append function for FEVector that adds a FEVectorBlock at the end.

Base.fill! — Method

fill!(b::FEVectorBlock, value::Any)

Custom fill function for FEVectorBlock (only fills the block, not the complete FEVector).

Base.show — Method

show(io::IO, FEF::FEVector)

Custom show function for FEVector that prints some information on its blocks.

GradientRobustMultiPhysics.addblock! — Method

addblock!(a::FEVectorBlock, b::AbstractArray{var"#s90",1} where var"#s90"<:Real; factor)

Adds Array b to FEVectorBlock a.

GradientRobustMultiPhysics.addblock! — Method

addblock!(a::FEVectorBlock, b::FEVectorBlock; factor)

Adds FEVectorBlock b to FEVectorBlock a.

FEMatrix

A FEMatrix consists of FEMatrixBlocks that share a common ExtendableSparseMatrix. Each block is associated to two FESpaces and can only write into a submatrix of the common sparse matrix specified by offsets. It also acts as a two-dimensional AbstractArray itself.

GradientRobustMultiPhysics.FEMatrix — Type

struct FEMatrix{T} <: AbstractArray{T,1}

an AbstractMatrix (e.g. an ExtendableSparseMatrix) with an additional layer of several FEMatrixBlock subdivisions each carrying coefficients for their associated pair of FESpaces

GradientRobustMultiPhysics.FEMatrix — Method

FEMatrix{T}(name::String, FES::Array{FESpace,1}) where T <: Real

Creates FEMatrix with blocks (FESX[i],FESY[j]) (enumerated row-wise).

GradientRobustMultiPhysics.FEMatrix — Method

FEMatrix{T}(name::String, FESX::FESpace, FESY::FESpace) where T <: Real

Creates FEMatrix with one rectangular block (FESX,FESY).

GradientRobustMultiPhysics.FEMatrix — Method

FEMatrix{T}(name::String, FES::FESpace) where T <: Real

Creates FEMatrix with one square block (FES,FES).

GradientRobustMultiPhysics.FEMatrixBlock — Type

struct FEMatrixBlock{T} <: AbstractArray{T,2}

block of an FEMatrix that carries coefficients for an associated pair of FESpaces and can be assigned as an two-dimensional AbstractArray (getindex, setindex, size)

Base.fill! — Method

fill!(B::FEMatrixBlock, value::Any)

Custom fill function for FEMatrixBlock (only fills the block, not the complete FEMatrix).

Base.show — Method

show(io::IO, FEM::FEMatrix)

Custom show function for FEMatrix that prints some information on its blocks.

GradientRobustMultiPhysics.addblock! — Method

addblock!(A::FEMatrixBlock, B::FEMatrixBlock; factor)

Adds FEMatrixBlock B to FEMatrixBlock A.

GradientRobustMultiPhysics.addblock! — Method

addblock!(A::FEMatrixBlock, B::ExtendableSparseMatrix{Tv,Ti<:Integer}; factor)

Adds ExtendableSparseMatrix B to FEMatrixBlock A.

GradientRobustMultiPhysics.addblock_matmul! — Method

addblock_matmul!(a::FEVectorBlock, B::FEMatrixBlock, b::FEVectorBlock; factor)

Adds matrix-vector product B times b to FEVectorBlock a.

GradientRobustMultiPhysics.addblock_matmul! — Method

addblock_matmul!(a::FEVectorBlock, B::ExtendableSparseMatrix{Tv,Ti<:Integer}, b::FEVectorBlock; factor)

Adds matrix-vector product B times b to FEVectorBlock a.

GradientRobustMultiPhysics.lrmatmul — Method

lrmatmul(a::AbstractArray{var"#s86",1} where var"#s86"<:Real, B::ExtendableSparseMatrix{Tv,Ti<:Integer}, b::AbstractArray{var"#s85",1} where var"#s85"<:Real; factor) -> Any

Computes vector'-matrix-vector product a'Bb.