Function Operators

FunctionOperators are building blocks for the weak form and define the operations that should be applied to the trial and test functions (and their discrete representatives) inside some PDEOperator. Below is a list of currently available FunctionOperators. Note, that not all operators an be applied to all finite element types.

Function operator	Description

Identity	identity

IdentityComponent{c}	identity of c-th component

IdentityDisc{Jump}	jumps of identity (only over faces)

NormalFlux	normal flux (function times normal)

Gradient	gradient/Jacobian

GradientDisc{Jump}	jumps of gradient/Jacobian (only over faces)

SymmetricGradient	symmetric part of the gradient

Divergence	divergence

CurlScalar	curl operator 1D to 2D (rotated gradient)

Curl2D	curl operator 2D to 1D

ReconstructionIdentity{FEType}	reconstruction operator into specified FEType

ReconstructionDivergence{FEType}	divergence of FEType reconstruction operator

Note

Especially note the operators ReconstructionIdentity{FEType} and ReconstructionDivergence{FEType} that allow to evaluate some reconstructed version of a vector-valued testfunction that maps its discrete divergence to the divergence and so allows e.g. gradient-robust discretisations with classical non divergence-conforming ansatz spaces. So far such operators are available for the vector-valued Crouzeix-Raviart and Bernardi–Raugel finite element types.

GradientRobustMultiPhysics.AbstractFunctionOperator — Type

abstract type AbstractFunctionOperator

root type for FunctionOperators.

GradientRobustMultiPhysics.Curl2D — Type

abstract type Curl2D <: AbstractFunctionOperator

evaluates the curl of some two-dimensional vector field, i.e. Curl2D((u1,u2)) = du2/dx1 - du1/dx2

GradientRobustMultiPhysics.CurlScalar — Type

abstract type CurlScalar <: AbstractFunctionOperator

evaluates the curl of some scalar function in 2D, i.e. the rotated gradient.

GradientRobustMultiPhysics.Divergence — Type

abstract type Divergence <: AbstractFunctionOperator

evaluates the divergence of the finite element function.

GradientRobustMultiPhysics.Gradient — Type

abstract type Gradient <: AbstractFunctionOperator

evaluates the gradient of the finite element function.

GradientRobustMultiPhysics.Identity — Type

abstract type Identity <: AbstractFunctionOperator

identity operator: evaluates finite element function.

GradientRobustMultiPhysics.IdentityComponent — Type

abstract type IdentityComponent <: AbstractFunctionOperator

identity operator: evaluates only the c-th component of the finite element function.

GradientRobustMultiPhysics.IdentityDisc — Type

abstract type IdentityDisc <: Identity

identity jump operator: evaluates face jumps of finite element function

GradientRobustMultiPhysics.NormalFlux — Type

abstract type NormalFlux <: AbstractFunctionOperator

evaluates the normal-flux of the finite element function.

only available on FACES/BFACES and currently only for H1 and Hdiv elements

GradientRobustMultiPhysics.ReconstructionDivergence — Type

abstract type ReconstructionDivergence <: Divergence

evaluates the divergence of the reconstructed finite element function.

FEreconst specifies the reconstruction space and reconstruction algorithm if it is defined for the finite element that it is applied to.

GradientRobustMultiPhysics.ReconstructionIdentity — Type

abstract type ReconstructionIdentity <: Identity

reconstruction identity operator: evaluates a reconstructed version of the finite element function.

FEreconst specifies the reconstruction space and reconstruction algorithm if it is defined for the finite element that it is applied to.

GradientRobustMultiPhysics.SymmetricGradient — Type

abstract type SymmetricGradient <: AbstractFunctionOperator

evaluates the symmetric part of the gradient of the finite element function.

GradientRobustMultiPhysics.TangentialGradient — Type

abstract type TangentialGradient <: AbstractFunctionOperator

evaluates the gradient of the tangential part of some vector-valued finite element function.