Item Integrators

Item integrators are one of the Assembly Patterns that help to compute certain quantities of the Solution, like a posteriori errors estimators, norms, drag/lift coefficients or other statistics.

GradientRobustMultiPhysics.ItemIntegrator — Function

function ItemIntegrator(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    name = "ItemIntegrator",
    regions::Array{Int,1} = [0])

Creates an ItemIntegrator assembly pattern with the given operators and action etc.

GradientRobustMultiPhysics.L2ErrorIntegrator — Function

function L2ErrorIntegrator(
    T::Type{<:Real},
    compare_data::UserData{AbstractDataFunction}, # can be omitted if zero
    operator::Type{<:AbstractFunctionOperator};
    quadorder = "auto",
    name = "auto",
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    time = 0)

Creates an ItemIntegrator that compares FEVectorBlock operator-evaluations against the given compare_data and returns the L2-error.

GradientRobustMultiPhysics.L2NormIntegrator — Function

L2NormIntegrator(
    T::Type{<:Real},
    ncomponents::Int,
    operator::Type{<:AbstractFunctionOperator};
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    name = "L2 norm",
    quadorder = 2,
    regions = [0])

Creates an ItemIntegrator that computes the L2 norm of an operator evaluation where ncomponents is the expected length of the operator evaluation.

GradientRobustMultiPhysics.L2DifferenceIntegrator — Function

function L2DifferenceIntegrator(
    T::Type{<:Real},
    ncomponents::Int,
    operator::Union{Type{<:AbstractFunctionOperator},Array{DataType,1}};
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    name = "L2 difference",
    quadorder = 2,
    regions = [0])

Creates an ItemIntegrator that computes the L2 norm difference between two arguments evalauted with the same operator (or with different operators if operator is an array) where ncomponents is the expected length of each operator evaluation. Note that all arguments in an evaluation call need to be defined on the same grid !

Evaluation

There are two possibilities to evaluate an ItemIntegrator, on each item (with evaluate!) or globally (with evaluate):

GradientRobustMultiPhysics.evaluate! — Function

function evaluate!(
    b::AbstractArray{T,2},          # target vector
    AP::AssemblyPattern{APT,T,AT},  # ItemIntegrator pattern
    FEB::Array{<:FEVectorBlock,1}   # coefficients for arguments
    where {APT <: APT_ItemIntegrator, T, AT}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlocks FEB into given two-dimensional Array b.

GradientRobustMultiPhysics.evaluate — Function

function evaluate(
    AP::AssemblyPattern{APT,T,AT},  # ItemIntegrator pattern
    FEB::Array{<:FEVectorBlock,1})  # coefficients for arguments
    where {APT <: APT_ItemIntegrator, T, AT}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlocks FEB, only returns accumulation over all items.

Noteworthy Examples

Examples 204 and A06 use ItemIntegrators for a posteriori error estimation and refinement indicators.

Example 224 uses ItemIntegrators to calculate drag and lift coefficients.