PDE Solvers

Fixed-Time Solvers

If solve! is applied to a PDEDescription and a FEVector (that specifies the ansatz spaces for the unknowns) an investigation of the PDEDescription is performed that decides if the problem is nonlinear (and has to be solved by a fixed-point algorithm) or if it can be solved directly in one step. Additionally the user can manually trigger subiterations that splits the fixed-point algorithm into substeps where only subsets of the PDE equations are solved together.

GradientRobustMultiPhysics.solve! — Function

function solve!(
    Target::FEVector,
    PDE::PDEDescription;
    kwargs)

Solves a given PDE (provided as a PDEDescription) and writes the solution into the FEVector Target (which knows the discrete ansatz spaces).

Keyword arguments:

anderson_iterations: use Anderson acceleration with this many previous iterates (to hopefully speed up/enable convergence of fixpoint iterations). Default: 0

subiterations: an array of equation subsets (each an array) that should be solved together in each fixpoint iteration. Default: auto

showiterationdetails: show details (residuals etc.) of each iteration. Default: true

anderson_unknowns: an array of unknown numbers that should be included in the Anderson acceleration. Default: [1]

show_statistics: show some statistics like assembly times. Default: false

anderson_metric: String that encodes the desired convergence metric for the Anderson acceleration (possible values: ''l2'' or ''L2''). Default: l2

skipupdate: matrix update (for the j-th sub-iteration) will be performed each skipupdate[j] iteration; -1 means only in the first iteration. Default: [1]

linsolver: String that encodes the linear solver (or type name of self-defined solver, see corressponding example). Default: UMFPACK

damping: damp the new iteration with this part of the old iteration. Default: 0

time: time at which time-dependent data functions are evaluated or initial time for TimeControlSolver. Default: 0

showsolverconfig: show the complete solver configuration before starting to solve. Default: false

checknonlinearresidual: check the nonlinear residual in last nonlinear iteration (causes one more reassembly of nonlinear terms). Default: auto

fixed_penalty: penalty that is used for the values of fixed degrees of freedom (e.g. by Dirichlet boundary data or global constraints). Default: 1.0e60

target_residual: stop fixpoint iterations if the (nonlinear) residual is smaller than this number. Default: 1.0e-10

maxiterations: maximal number of nonlinear iterations (TimeControlSolver runs that many in each time step). Default: auto

Depending on the subiterations and detected/configured nonlinearities the whole system is either solved directly in one step or via a fixed-point iteration.

Time-Dependent Solvers

The structure TimeControlSolver can be used to setup a time-dependent solver that can be configured in a similar manner as the time-independent ones (subiterations, nonlinear iterations, linear solvers). As a TimeIntegrationRule so far only BackwardEuler is implemented.

Note that, the time-derivative is added by the TimeControlSolver and is in general not part of the PDEDescription (this is debatable and might change in the future).

GradientRobustMultiPhysics.TimeControlSolver — Type

function TimeControlSolver(
    PDE::PDEDescription,
    Solution::FEVector,    # contains initial values and stores solution of advance method below
    TIR::Type{<:AbstractTimeIntegrationRule} = BackwardEuler;
    dt_testfunction_operator = [],
    dt_action = [],
    nonlinear_dt = [],
    kwargs...)

Creates a time-dependent solver that can be advanced in time with advance!. The FEVector Solution stores the initial state but also the solution at the current time. The argument TIR carries the time integration rule (currently there is only BackwardEuler).

Keyword arguments:

subiterations: an array of equation subsets (each an array) that should be solved together in each fixpoint iteration. Default: auto

showiterationdetails: show details (residuals etc.) of each iteration. Default: true

timedependent_equations: array of the equations that should get a time derivative (only for TimeControlSolver). Default: Any[]

show_statistics: show some statistics like assembly times. Default: false

skipupdate: matrix update (for the j-th sub-iteration) will be performed each skipupdate[j] iteration; -1 means only in the first iteration. Default: [1]

linsolver: String that encodes the linear solver (or type name of self-defined solver, see corressponding example). Default: UMFPACK

time: time at which time-dependent data functions are evaluated or initial time for TimeControlSolver. Default: 0

showsolverconfig: show the complete solver configuration before starting to solve. Default: false

checknonlinearresidual: check the nonlinear residual in last nonlinear iteration (causes one more reassembly of nonlinear terms). Default: auto

fixed_penalty: penalty that is used for the values of fixed degrees of freedom (e.g. by Dirichlet boundary data or global constraints). Default: 1.0e60

target_residual: stop fixpoint iterations if the (nonlinear) residual is smaller than this number. Default: 1.0e-10

maxiterations: maximal number of nonlinear iterations (TimeControlSolver runs that many in each time step). Default: auto

Further (very experimental) optional arguments for TimeControlSolver are:

dttestfunction_operator : (array of) operators applied to testfunctions in time derivative (default: Identity)

dt_action : (array of) actions that are applied to the ansatz function in the time derivative (to include parameters etc.)

nonlinear_dt : (array of) booleans to decide which time derivatives should be recomputed in each iteration/timestep

GradientRobustMultiPhysics.advance! — Function

function TimeControlSolver(
    advance!(TCS::TimeControlSolver, timestep::Real = 1e-1)

Advances a TimeControlSolver one step in time with the given timestep.

Moreover there are two functions that advance the TimeControlSolver automatically until a given final time (advance_until_time!) is reached or until stationarity is reached (advance_until_stationarity!). As an experimental feature, one can add the module DifferentialEquations.jl as the first argument to these methods to let this module run the time integration.

GradientRobustMultiPhysics.advance_until_time! — Function

advance_until_time!(TCS::TimeControlSolver, timestep, finaltime; finaltime_tolerance = 1e-15, do_after_each_timestep = nothing)

Advances a TimeControlSolver in time with the given (initial) timestep until the specified finaltime is reached (up to the specified tolerance). The function doaftertimestep is called after each timestep and can be used to print/save data (and maybe timestep control in future).

advance_until_time!(DiffEQ::Module, TCS::TimeControlSolver, timestep, finaltime; solver = nothing, abstol = 1e-1, reltol = 1e-1, dtmin = 0, adaptive::Bool = true)

Advances a TimeControlSolver in time with the given (initial) timestep until the specified finaltime is reached (up to the specified tolerance) with the given exterior time integration module. The only valid Module here is DifferentialEquations.jl and the optional arguments are passed to it. If solver == nothing the solver Rosenbrock23(autodiff = false) will be chosen. For more choices please consult the documentation of DifferentialEquations.jl.

Also note that this is a highly experimental feature and will not work for general TimeControlSolvers configuration (e.g. in the case of several subiterations or, it seems, saddle point problems). Also have a look at corressponding the example in the advanced examples section.

GradientRobustMultiPhysics.advance_until_stationarity! — Function

advance_until_stationarity!(TCS::TimeControlSolver, timestep; stationarity_threshold = 1e-11, maxTimeSteps = 100, do_after_each_timestep = nothing)

Advances a TimeControlSolver in time with the given (initial) timestep until stationarity is detected (change of variables below threshold) or a maximal number of time steps is exceeded. The function doaftertimestep is called after each timestep and can be used to print/save data (and maybe timestep control in future).