Export and Viewers

Plotting is possible e.g. via calculation of nodal values and the plot routines from ExtendableGrids.jl. It is also possible to export (the nodal values) of a FEVector as a vtk datafile and load it into e.g. paraview.

GradientRobustMultiPhysics.plot — Method

plot(Source::FEVector, blockids::Array{Int64,1}, operators::Array{DataType,1}; Plotter, use_subplots, subplots_per_column, colorlevels, isolines, aspect, show, clear, cbar, verbosity, cmap, maintitle, fsize)

Plots operators applied to components of a given FEVector Source (operator[j9] is applied to component with blockids[j]) via ExtendablePlots.plot, either in one window with subplots or several single plots.

If blockids[j] == 0, a plot of the grid is generated at the j-th subplot.

Plotter = PyPlot should work in 2D Plotter = Makie should work in 3D (but only with a single plot currently)

GradientRobustMultiPhysics.writeCSV! — Method

writeCSV!(filename::String, Source::FEVector; blockids, operators, names, seperator, verbosity)

Writes the specified FEVector into a CSV datafile with the given filename. First d colomuns are the grid coordinates, the remaining column are filled with the evaluations of the operators where operator[j] is applied to Source[blockids[j]].

GradientRobustMultiPhysics.writeVTK! — Method

writeVTK!(filename::String, Data::FEVector; blocks, operators, names, vectorabs)

Writes the specified FEVector into a vtk datafile with the given filename. Each component of each FEVectorBlock (or the subset specified by blocks) is saved separately. Vector-valued quantities also generate a data field that represents the absolute value of the vector field at each grid point.