Assembly Details

The assembly of an operator is essentially based on a combination of Assembly Types and Assembly Patterns. The assembly type mainly allows to choose the geometry information needed for providing quadrature and dof handling. The assembly pattern then basically evaluates the function operators and action for the ansatz and test functions and does the quadrature-weighted accumulation into matrices or vectors that represent the operators.

Assembly Types

The following assembly types are available.

AssemblyType	Description

ON_CELLS	assembles over the cells of the mesh

ON_FACES	assembles over all faces of the mesh

ON_IFACES	assembles over the interior faces of the mesh

ON_BFACES	assembles over the boundary faces of the mesh

ON_EDGES (*)	assembles over all edges of the mesh (in 3D)

ON_BEDGES (*)	assembles over the boundary edges of the mesh (in 3D)

Note

(*) = only reasonable in 3D and still experimental, might have some issues

Assembly Patterns

Each Pattern comes with a number of arguments/quantities with associated Function Operators as well as one of the Assembly Types that states whether the form is evaluated over CELLS, FACES order BFACES (see above). Important note: this assembly type is relative to the grid of the first argument of the pattern. If this argument already lives ONFACES and the pattern is also ONFACES, it will ultimatively assemble on the faces of the faces (that are the edges of the grid with these faces). Moreover, patterns can have Abstract Actions that allow to make the evaluations parameter-, region- and/or function-dependent. Each pattern then has usually on to three implementation that writes into FEMatrix or FEVector (where e.g. a subset of arguments is fixed) or evaluates the pattern in the given FEVectorBlocks.

The patterns are used by the assembly of the PDE operators defined in a PDE Description. However, it is also possible for the user to use them directly, see e.g. the example Commuting Interpolators (2D).

The following table lists all available assembly patterns and how they can be used for assembly or evaluations.

AssemblyPattern	evaluate	assemble into matrix	assembled into vector

ItemIntegrator	yes	no	no

LinearForm	no	no	yes

BilinearForm	no	yes	yes (1)

TrilinearForm	no	yes (1)	yes (2)

MultilinearForm	no	no	yes (N-1)

NonlinearForm	no	yes (L)	yes (L)

Number in brackets denotes the number of fixed arguments needed for this assembly, (L) means that a current solution is needed to evaluate (to evaluate the linearisation of the nonlinear form in this state). Evaluations of the other AssemblyPatterns will be possible in a future update, but currently have to be performed by maintaining a duplicate of the pattern rewritten as an ItemIntegrator.

ItemIntegrator

GradientRobustMultiPhysics.ItemIntegrator — Function

function ItemIntegrator(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates an ItemIntegrator assembly pattern with the given operators and action etc.

Linearform

GradientRobustMultiPhysics.LinearForm — Function

function LinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a LinearForm assembly pattern with the given FESpaces, operators and action etc.

Bilinearform

GradientRobustMultiPhysics.BilinearForm — Function

function BilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a (unsymmetric) BilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.SymmetricBilinearForm — Function

function SymmetricBilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a symmetric BilinearForm assembly pattern with the given FESpaces, operators and action etc. Symmetry is not checked automatically, but is assumed during assembly!

Trilinearform

TrilinearForm

Multilinearform

GradientRobustMultiPhysics.MultilinearForm — Function

function MultilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a MultilinearForm assembly pattern with the given FESpaces, operators and action etc.

Nonlinearform

GradientRobustMultiPhysics.NonlinearForm — Function

function NonlinearForm(
    T::Type{<:Real},
    FES::Array{FESpace,1},          # finite element spaces for each operator of the ansatz function and the last one refers to the test function
    operators::Array{DataType,1},   # operators that should be evaluated for the ansatz function and the last one refers to the test function
    action::AbstractAction;         # action that shoul have an AbstractNLActionKernel
    regions::Array{Int,1} = [0])

Creates a NonlinearForm assembly pattern.

Evaluate! & Assemble!

GradientRobustMultiPhysics.evaluate! — Function

function evaluate!(
    b::AbstractArray{T,2},
    AP::AssemblyPattern{APT,T,AT},
    FEB::FEVectorBlock) where {APT <: APT_ItemIntegrator, T<: Real, AT <: AbstractAssemblyType}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlocks FEB into given two-dimensional Array b.

GradientRobustMultiPhysics.assemble! — Function

assemble!(
    A::AbstractArray{T,2},
    AP::AssemblyPattern{APT,T,AT};
    apply_action_to::Int = 1,
    factor = 1,
    transposed_assembly::Bool = false,
    transpose_copy = Nothing)  where {APT <: APT_BilinearForm, T <: Real, AT <: AbstractAssemblyType}

Assembly of a BilinearForm BLF into given two-dimensional AbstractArray (e.g. FEMatrixBlock).

assemble!(
    b::AbstractArray{T,1},
    fixedFE::FEVectorBlock,    # coefficient for fixed 2nd component
    AP::AssemblyPattern{APT,T,AT};
    apply_action_to::Int = 1,
    fixed_argument::Int = 2,
    factor = 1) where where {APT <: APT_BilinearForm, T <: Real, AT <: AbstractAssemblyType}

Assembly of a BilinearForm BLF into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the second argument is fixed (default) by the given coefficients in fixedFE. With applyactionto=2 the action can be also applied to the second argument instead of the first one (default).

assemble!(
    assemble!(
    A::AbstractArray{T,2},
    FE1::FEVectorBlock,
    AP::TrilinearForm{T, AT};
    fixed_argument::Int = 1,
    transposed_assembly::Bool = false,
    factor = 1)

Assembly of a TrilinearForm AP into given two-dimensional AbstractArray (e.g. a FEMatrixBlock). Here, one argument (specified by fixed_argument) is fixed by the given coefficients in FE1. Note, that the action is (currently) always applied to the first and second argument.

assemble!(
    assemble!(
    b::AbstractVector,
    FE1::FEVectorBlock,
    FE2::FEVectorBlock.
    AP::AssemblyPattern{APT,T,AT};
    factor = 1)

Assembly of a TrilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the first two arguments are fixed by the given coefficients in FE1 and FE2.

assemble!(
    assemble!(
    b::AbstractVector,
    FE::Array{<:FEVectorBlock,1},
    AP::AssemblyPattern{APT,T,AT};
    factor = 1)

Assembly of a MultilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the all but the last arguments are fixed by the given coefficients in the components of FE.

assemble!(
    A::AbstractArray{T,2},
    AP::AssemblyPattern{APT,T,AT};
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    factor = 1,
    transposed_assembly::Bool = false

Assembly of a NonlinearForm assembly pattern into given two-dimensional AbstractArray (e.g. FEMatrixBlock).

assemble!(
    b::AbstractVector,
    AP::NonlinearForm{T, AT},
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    factor = 1,
    transposed_assembly::Bool = false

Assembly of a NonlinearForm AP into given AbstractVector (e.g. FEMatrixBlock).