Assembly Details

The assembly of an operator is essentially based on a combination of Assembly Types and Assembly Patterns. The assembly type mainly allows to choose the geometry information needed for providing quadrature and dof handling. The assembly pattern then basically evaluates the function operators and action for the ansatz and test functions and does the quadrature-weighted accumulation into matrices or vectors that represent the operators.

Assembly Types

The following assembly types are available.

AssemblyType	Description

ON_CELLS	assembles over the cells of the mesh

ON_FACES	assembles over all faces of the mesh

ON_IFACES (*)	assembles over the interior faces of the mesh

ON_BFACES	assembles over the boundary faces of the mesh

ON_EDGES (*)	assembles over all edges of the mesh (in 3D)

ON_BEDGES (*)	assembles over the boundary edges of the mesh (in 3D)

Note

(*) = still experimental, might have some issues

Assembly Patterns

Each Pattern comes with a number of arguments/quantities with associated Function Operators as well as one of the Assembly Types that states whether the form is evaluated over CELLS, FACES order BFACES (see above). Moreover, patterns can have Abstract Actions that allow to make the evaluations parameter-, region- and/or function-dependent. Each pattern then has two implementation that writes into FEMatrix or FEVector (where e.g. a subset of arguments is fixed).

The patterns are used by the assembly of the PDE operators defined in a PDE Description. However, it is also possible for the user to use them directly, see e.g. the example Commuting Interpolators (2D).

GradientRobustMultiPhysics.AbstractAssemblyPattern — Type

abstract type AbstractAssemblyPattern

abstract type for assembly patterns

GradientRobustMultiPhysics.BilinearForm — Type

struct BilinearForm{T<:Real, AT<:AbstractAssemblyType} <: AbstractAssemblyPattern{T<:Real,AT<:AbstractAssemblyType}

assembly pattern bilinear form (that depends on two quantities)

GradientRobustMultiPhysics.BilinearForm — Method

function BilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE1::FESpace,
    FE2::FESpace,
    operator1::Type{<:AbstractFunctionOperator},
    operator2::Type{<:AbstractFunctionOperator},
    action::AbstractAction; # is only applied to FE1/operator1
    regions::Array{Int,1} = [0])

Creates an unsymmetric BilinearForm.

GradientRobustMultiPhysics.ItemIntegrator — Type

struct ItemIntegrator{T<:Real, AT<:AbstractAssemblyType} <: AbstractAssemblyPattern{T<:Real,AT<:AbstractAssemblyType}

assembly pattern item integrator that can e.g. be used for error/norm evaluations

GradientRobustMultiPhysics.LinearForm — Type

struct LinearForm{T<:Real, AT<:AbstractAssemblyType} <: AbstractAssemblyPattern{T<:Real,AT<:AbstractAssemblyType}

assembly pattern linear form (that only depends on one quantity)

GradientRobustMultiPhysics.LinearForm — Method

function LinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::FESpace,
    operator::Type{<:AbstractFunctionOperator},
    action::AbstractAction;
    regions::Array{Int,1} = [0])

Creates a LinearForm.

GradientRobustMultiPhysics.MultilinearForm — Type

struct MultilinearForm{T<:Real, AT<:AbstractAssemblyType} <: AbstractAssemblyPattern{T<:Real,AT<:AbstractAssemblyType}

assembly pattern multi-linear form (that depends on arbitrary many quantities)

currently this can be only assembled into a FEVector (with one free argument, all others fixed)

GradientRobustMultiPhysics.MultilinearForm — Method

function MultilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a MultilinearForm that can be only assembled into a vector (with all but one fixed arguments).

GradientRobustMultiPhysics.NonlinearForm — Type

struct NonlinearForm{T<:Real, AT<:AbstractAssemblyType} <: AbstractAssemblyPattern{T<:Real,AT<:AbstractAssemblyType}

assembly pattern nonlinear form with two arguments (ansatz and testfunction) where te first argument can depend on more than one operator

GradientRobustMultiPhysics.NonlinearForm — Method

function NonlinearForm(
    T::Type{<:Real},
    FE1::Array{FESpace,1},          # finite element spaces for each operator of the ansatz function
    FE2::FESpace,                   # finite element space for testfunction
    operator1::Array{DataType,1},   # operators that should be evaluated for the ansatz function
    operator2::Type{<:AbstractFunctionOperator},   # operator that is evaluated for the test function
    action::AbstractAction;        # is applied to all operators of ansatz functions (to allow for Newton)
                                   # in Newton mode also the evaluation of all operators in current solution are
                                   # passed to the action
    ADnewton::Bool = false,        # perform AD to obtain additional Newton terms
    regions::Array{Int,1} = [0])

Creates a NonlinearForm.

GradientRobustMultiPhysics.TrilinearForm — Type

struct TrilinearForm{T<:Real, AT<:AbstractAssemblyType} <: AbstractAssemblyPattern{T<:Real,AT<:AbstractAssemblyType}

assembly pattern trilinear form (that depends on three quantities)

GradientRobustMultiPhysics.TrilinearForm — Method

function TrilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE1::FESpace,
    FE2::FESpace,
    FE3::FESpace,
    operator1::Type{<:AbstractFunctionOperator},
    operator2::Type{<:AbstractFunctionOperator},
    operator3::Type{<:AbstractFunctionOperator},
    action::AbstractAction; # is only applied to FE1/operator1 + FE2/operator2
    regions::Array{Int,1} = [0])

Creates a TrilinearForm that can be assembeld into a matrix (with one argument fixed) or into a vector (with two fixed arguments).

GradientRobustMultiPhysics.L2ErrorIntegrator — Method

function L2ErrorIntegrator(
    T::Type{<:Real},
    compare_data::UserData{AbstractDataFunction},
    operator::Type{<:AbstractFunctionOperator};
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    time = 0)

Creates an ItemIntegrator that compares FEVectorBlock operator-evaluations against the given compare_data and returns the L2-error.

GradientRobustMultiPhysics.SymmetricBilinearForm — Method

function SymmetricBilinearForm(
    T::Type{<:Real},
    AT::Type{<:AbstractAssemblyType},
    FE1::FESpace,
    operator1::Type{<:AbstractFunctionOperator},
    action::AbstractAction; # is only applied to FE1/operator1
    regions::Array{Int,1} = [0])

Creates a symmetric BilinearForm that can be assembled into a matrix or a vector (with one argument fixed)

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    assemble!(
    b::AbstractVector,
    FE::Array{<:FEVectorBlock,1},
    MLF::MultilinearForm{T, AT};
    verbosity::Int = 0,
    factor = 1)

Assembly of a MultilinearForm MLF into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the all but the last arguments are fixed by the given coefficients in the components of FE.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    assemble!(
    b::AbstractVector,
    FE1::FEVectorBlock,
    FE2::FEVectorBlock.
    TLF::TrilinearForm{T, AT};
    verbosity::Int = 0,
    factor = 1)

Assembly of a TrilinearForm TLF into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the first two arguments are fixed by the given coefficients in FE1 and FE2.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    b::AbstractVector,
    NLF::NonlinearForm{T, AT},
    FEB::Array{<:FEVectorBlock,1};         # coefficients for each operator
    verbosity::Int = 0,
    factor = 1,
    transposed_assembly::Bool = false

Assembly of a NonlinearForm NLF into given AbstractVector (e.g. FEMatrixBlock).

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    b::AbstractArray{T,1},
    fixedFE::FEVectorBlock,    # coefficient for fixed 2nd component
    BLF::BilinearForm{T, AT};
    apply_action_to::Int = 1,
    fixed_argument::Int = 2,
    factor = 1,
    verbosity::Int = 0) where {T<: Real, AT <: AbstractAssemblyType}

Assembly of a BilinearForm BLF into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the second argument is fixed (default) by the given coefficients in fixedFE. With applyactionto=2 the action can be also applied to the second argument instead of the first one (default).

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    A::AbstractArray{T,2},
    BLF::BilinearForm{T, AT};
    apply_action_to::Int = 1,
    verbosity::Int = 0,
    factor = 1,
    transposed_assembly::Bool = false,
    transpose_copy = Nothing) where {T<: Real, AT <: AbstractAssemblyType}

Assembly of a BilinearForm BLF into given two-dimensional AbstractArray (e.g. FEMatrixBlock).

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    assemble!(
    A::AbstractArray{T,2},
    FE1::FEVectorBlock,
    TLF::TrilinearForm{T, AT};
    verbosity::Int = 0,
    fixed_argument::Int = 1,
    transposed_assembly::Bool = false,
    factor = 1)

Assembly of a TrilinearForm TLF into given two-dimensional AbstractArray (e.g. a FEMatrixBlock). Here, the first argument is fixed by the given coefficients in FE1.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    A::AbstractArray{T,2},
    NLF::NonlinearForm{T, AT},
    FEB::Array{<:FEVectorBlock,1};         # coefficients for each operator
    verbosity::Int = 0,
    factor = 1,
    transposed_assembly::Bool = false

Assembly of a NonlinearForm NLF into given two-dimensional AbstractArray (e.g. FEMatrixBlock).

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    b::Union{AbstractArray{T,1},AbstractArray{T,2}},
    LF::LinearForm{T,AT};
    verbosity::Int = 0) where {T<: Real, AT <: AbstractAssemblyType}

Assembly of a LinearForm LF into given one- or two-dimensional Array b.

GradientRobustMultiPhysics.evaluate! — Method

function evaluate!(
    b::AbstractArray{T,2},
    form::ItemIntegrator{T,AT},
    FEB::FEVectorBlock;
    verbosity::Int = 0) where {T<: Real, AT <: AbstractAssemblyType}

Evaluation of an ItemIntegrator form with given FEVectorBlock FEB into given two-dimensional Array b.

GradientRobustMultiPhysics.evaluate — Method

function evaluate(
    form::ItemIntegrator{T,AT},
    FEB::FEVectorBlock;
    verbosity::Int = 0) where {T<: Real, AT <: AbstractAssemblyType}

Evaluation of an ItemIntegrator form with given FEVectorBlock FEB, only returns accumulation over all items.