Assembly Patterns

The definition and assembly of an operator is essentially based on Assembly Patterns and a Assembly Type to choose the geometry information needed for providing quadrature and dof handling. The assembly pattern then basically evaluates the function operators and action for the ansatz and test functions and does the quadrature-weighted accumulation into matrices or vectors that represent the operators.

Each pattern comes with a number of arguments/quantities with associated Function Operators as well as one of the Assembly Type that states whether the form is evaluated over CELLS, FACES order BFACES (see above). Important note: this assembly type is relative to the grid of the first argument of the pattern. If this argument already lives ONFACES and the pattern is also ONFACES, it will ultimatively assemble on the faces of the faces (that are the edges of the grid with these faces). Moreover, patterns can have an Action that allow to make the evaluations parameter-, region- and/or function-dependent. Each pattern then has usually on to three implementation that writes into FEMatrix or FEVector (where e.g. a subset of arguments is fixed) or evaluates the pattern in the given FEVectorBlocks.

The patterns are used to assembly the PDE operators defined in a PDE Description.

GradientRobustMultiPhysics.AssemblyPattern — Type

mutable struct AssemblyPattern{APT<:AssemblyPatternType, T<:Real, AT<:AssemblyType, Tv<:Real, Ti<:Integer, ActionType<:AbstractAction}

each assembly pattern has one of the assembly pattern types (APT) that trigger different assemblies for the involved finite element spaces, operators and an assigned action. The assembly type (AT) determines if the assembly takes place on cells, faces or edges etc. (relatively to the assembly type of the first argument of the pattern)

The following table lists all available assembly patterns, their constuctor names and how they can be used for assembly or evaluations.

AssemblyPatternType	constructor	evaluate	assembly into matrix	assembly into vector

APT_ItemIntegrator	ItemIntegrator	yes	no	no

APT_LinearForm	LinearForm	no	no	yes

APT_BilinearForm	BilinearForm	no	yes	yes (1)

APT_TrilinearForm	TrilinearForm	no	yes (1)	yes (2)

APT_MultiLinearForm	MultilinearForm	no	no	yes (N-1)

APT_NonlinearForm	NonlinearForm	no	yes (L)	yes (L)

Number in brackets denotes the number of fixed arguments needed for this assembly, (L) means that a current solution is needed to evaluate (to evaluate the linearisation of the nonlinear form in this state). Evaluations of the other AssemblyPatterns may be possible in a future update, but currently have to be performed by maintaining a duplicate of the pattern rewritten as an ItemIntegrator.

Constructor details

Below all assembly pattern types, constructor functions and evaluate/assembly functions are detailed. (For more on the ItemIntegrator also see Item Integrators.)

GradientRobustMultiPhysics.APT_ItemIntegrator — Type

abstract type APT_ItemIntegrator <: AssemblyPatternType

itemintegrator assembly pattern type

GradientRobustMultiPhysics.ItemIntegrator — Function

function ItemIntegrator(
    T::Type{<:Real},
    AT::Type{<:AssemblyType},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    name = "ItemIntegrator",
    regions::Array{Int,1} = [0])

Creates an ItemIntegrator assembly pattern with the given operators and action etc.

GradientRobustMultiPhysics.L2DifferenceIntegrator — Method

function L2DifferenceIntegrator(
    T::Type{<:Real},
    ncomponents::Int,
    operator::Union{Type{<:AbstractFunctionOperator},Array{DataType,1}};
    AT::Type{<:AssemblyType} = ON_CELLS,
    name = "L2 difference",
    quadorder = 2,
    regions = [0])

Creates an ItemIntegrator that computes the L2 norm difference between two arguments evalauted with the same operator (or with different operators if operator is an array) where ncomponents is the expected length of each operator evaluation. Note that all arguments in an evaluation call need to be defined on the same grid !

GradientRobustMultiPhysics.L2ErrorIntegrator — Function

function L2ErrorIntegrator(
    T::Type{<:Real},
    compare_data::UserData{AbstractDataFunction}, # can be omitted if zero
    operator::Type{<:AbstractFunctionOperator};
    quadorder = "auto",
    name = "auto",
    factor = 1,
    AT::Type{<:AssemblyType} = ON_CELLS,
    time = 0)

Creates an ItemIntegrator that compares discrete FEVectorBlock operator-evaluations against the given comparedata and returns the L2-error || comparedata(x) - factor*discrete(x) ||. If quadorder is left on "auto" two times the quadorder of the data is used in the evaluation.

GradientRobustMultiPhysics.L2NormIntegrator — Method

L2NormIntegrator(
    T::Type{<:Real},
    ncomponents::Int,
    operator::Type{<:AbstractFunctionOperator};
    AT::Type{<:AssemblyType} = ON_CELLS,
    name = "L2 norm",
    quadorder = 2,
    regions = [0])

Creates an ItemIntegrator that computes the L2 norm of an operator evaluation where ncomponents is the expected length of the operator evaluation.

GradientRobustMultiPhysics.evaluate! — Method

function evaluate!(
    b::AbstractArray{T,2},
    AP::AssemblyPattern{APT,T,AT},
    FEB::Union{<:FEVector{T,Tv,Ti},<:FEVectorBlock{T,Tv,Ti},Array{<:FEVectorBlock{T,Tv,Ti},1}};
    skip_preps::Bool = false) where {APT <: APT_ItemIntegrator, T<: Real, AT <: AssemblyType, Tv, Ti}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlock or FEVector FEB into given two-dimensional Array b.

GradientRobustMultiPhysics.evaluate — Method

function evaluate(
    AP::AssemblyPattern{APT,T,AT},
    FEB::Union{<:FEVector{T,Tv,Ti},<:FEVectorBlock{T,Tv,Ti},Array{<:FEVectorBlock{T,Tv,Ti},1}};
    skip_preps::Bool = false) where {APT <: APT_ItemIntegrator, T<: Real, AT <: AssemblyType, Tv, Ti}

Evaluation of an ItemIntegrator assembly pattern with given FEVectorBlock or FEVector FEB, only returns accumulation over all items.

GradientRobustMultiPhysics.APT_BilinearForm — Type

abstract type APT_BilinearForm <: AssemblyPatternType

bilinearform assembly pattern type

GradientRobustMultiPhysics.APT_LumpedBilinearForm — Type

abstract type APT_LumpedBilinearForm <: BilinearForm

lumped bilinearform assembly pattern type where only the diagonal elements on each item are assembled

GradientRobustMultiPhysics.APT_SymmetricBilinearForm — Type

abstract type APT_SymmetricBilinearForm <: BilinearForm

symmetric bilinearform assembly pattern type

GradientRobustMultiPhysics.BilinearForm — Function

function BilinearForm(
    T::Type{<:Real},
    AT::Type{<:AssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a general BilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.LumpedBilinearForm — Function

function LumpedBilinearForm(
    T::Type{<:Real},
    AT::Type{<:AssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a LumpedBilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.SymmetricBilinearForm — Function

function SymmetricBilinearForm(
    T::Type{<:Real},
    AT::Type{<:AssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a symmetric BilinearForm assembly pattern with the given FESpaces, operators and action etc. Symmetry is not checked automatically, but is assumed during assembly!

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    b::AbstractArray{T,1},          # target vector
    AP::AssemblyPattern{APT,T,AT},  # BilinearForm Pattern
    fixedFE::AbstractArray;         # coefficients for fixed argument
    apply_action_to::Int = 1,       # action is applied to 1st or 2nd argument?
    fixed_arguments = [1],        # which argument is fixed?
    factor = 1)                     # factor that is multiplied
    where {APT <: APT_BilinearForm, T, AT}

Assembly of a BilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the second argument is fixed (default) by the given coefficients in fixedFE. With applyactionto=2 the action can be also applied to the second argument instead of the first one (default).

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    A::AbstractArray{T,2},                  # target matrix
    AP::AssemblyPattern{APT,T,AT};          # BilinearForm Pattern
    apply_action_to::Int = 1,               # action is applied to which argument?
    factor = 1,                             # factor that is multiplied
    transposed_assembly::Bool = false,      # transpose result?
    transpose_copy = Nothing)               # copy a transposed block to this matrix
    where {APT <: APT_BilinearForm, T, AT}

Assembly of a BilinearForm BLF into given two-dimensional AbstractArray (e.g. FEMatrixBlock or a ExtendableSparseMatrix).

GradientRobustMultiPhysics.APT_TrilinearForm — Type

abstract type APT_TrilinearForm <: AssemblyPatternType

trilinearform assembly pattern type

GradientRobustMultiPhysics.TrilinearForm — Method

function TrilinearForm(
    T::Type{<:Real},
    FES::Array{FESpace,1},          
    operators::Array{DataType,1},
    action::AbstractAction;
    regions::Array{Int,1} = [0])

Creates a TrilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    assemble!(
    A::AbstractArray{T,2},                  # target matrix
    AP::AssemblyPattern{APT,T,AT},          # TrilinearForm pattern
    fixedFE::Array{<:FEVectorBlock,1};      # coefficients for fixed argument
    fixed_argument = [1],                   # position of fixed argument (as an array)
    transposed_assembly::Bool = false,      # transpose result?
    factor = 1)                             # factor that is multiplied
    where {APT <: APT_TrilinearForm, T, AT}

Assembly of a TrilinearForm AP into given two-dimensional AbstractArray (e.g. a FEMatrixBlock). Here, one argument (specified by fixed_argument) is fixed by the given coefficients in fixedFE[1]. Note, that the action is (currently) always applied to the first and second argument.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    assemble!(
    b::AbstractVector,                      # target vector
    AP::AssemblyPattern{APT,T,AT},          # coefficients for fixed arguments
    fixedFE::Array{<:FEVectorBlock,1};      # TrilinearForm pattern
    factor = 1)                             # factor that is multiplied
    where {APT <: APT_TrilinearForm, T, AT}

Assembly of a TrilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the first two arguments are fixed by the given coefficients in FE1 and FE2.

GradientRobustMultiPhysics.APT_MultilinearForm — Type

abstract type APT_MultilinearForm <: AssemblyPatternType

multilinearform assembly pattern type

GradientRobustMultiPhysics.MultilinearForm — Method

function MultilinearForm(
    T::Type{<:Real},
    AT::Type{<:AssemblyType},
    FE::Array{FESpace,1},
    operators::Array{DataType,1}, 
    action::AbstractAction; 
    regions::Array{Int,1} = [0])

Creates a MultilinearForm assembly pattern with the given FESpaces, operators and action etc.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    b::AbstractVector,                        # target vector
    FE::Array{<:FEVectorBlock,1},             # coefficients of all but last argument
    AP::AssemblyPattern{APT,T,AT};            # Multilinearform pattern
    factor = 1)                               # factor that is multiplied
    where {APT <: APT_MultilinearForm, T, AT}

Assembly of a MultilinearForm AP into given one-dimensional AbstractArray (e.g. a FEVectorBlock). Here, the all but the last arguments are fixed by the given coefficients in the components of FE.

GradientRobustMultiPhysics.APT_NonlinearForm — Type

abstract type APT_NonlinearForm <: AssemblyPatternType

nonlinearform assembly pattern type

GradientRobustMultiPhysics.NonlinearForm — Method

function NonlinearForm(
    T::Type{<:Real},
    FES::Array{FESpace,1},          # finite element spaces for each operator of the ansatz function and the last one refers to the test function
    operators::Array{DataType,1},   # operators that should be evaluated for the ansatz function and the last one refers to the test function
    action::AbstractAction;         # action that shoul have an AbstractNLActionKernel
    regions::Array{Int,1} = [0])

Creates a NonlinearForm assembly pattern.

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    A::AbstractArray{T,2},                 # target matrix
    AP::AssemblyPattern{APT,T,AT};         # NonlinearForm pattern
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    factor = 1,                            # factor that is multiplied
    transposed_assembly::Bool = false)     # transpose result?
    where {APT <: APT_NonlinearForm, T, AT}

Assembly of a NonlinearForm assembly pattern into given two-dimensional AbstractArray (e.g. FEMatrixBlock).

GradientRobustMultiPhysics.assemble! — Method

assemble!(
    b::AbstractVector,                     # target vector
    AP::AssemblyPattern{APT,T,AT},         # NonlinearForm pattern
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    factor = 1)                            # factor that is multiplied
    where {APT <: APT_NonlinearForm, T, AT}

Assembly of a NonlinearForm AP into given AbstractVector (e.g. FEMatrixBlock).

GradientRobustMultiPhysics.evaluate — Method

assemble!(
    AP::AssemblyPattern{APT,T,AT},         # NonlinearForm pattern
    FEB::Array{<:FEVectorBlock,1};         # coefficients of current solution for each operator
    FEBtest::FEVectorBlock;                # coefficients of test function for test function operator
    factor = 1)                            # factor that is multiplied
    where {APT <: APT_NonlinearForm, T, AT}

Evaluation of a NonlinearForm AP for given coefficients of ansatz and test function.