Data Export

Here some export functionality is documented. Also note that ExtendableGrids.jl offers an interface to WriteVTK.jl which can be used in combination with nodevalues interpolations.

GradientRobustMultiPhysics.writeCSV! — Method

writeCSV!(filename::String, Data::Vector{<:FEVectorBlock}; operators, names, seperator)

Writes the specified FEVectorBlocks into a CSV datafile with the given filename. First d colomuns are the grid coordinates, the remaining columns are filled with the evaluations of the operators where operator[j] is applied to Data[j].