Implemented Finite Elements

This page describes the finite element type-tree and lists all implemented finite elements.

The Finite Element Type-Tree

Finite elements are abstract type leaves in a type-tree. Its root and first layer look like this

AbstractFiniteElements

AbstractH1FiniteElement

AbstractHdivFiniteElement

AbstractHcurlFiniteElement

AbstractL2FiniteElement

Remarks:

each finite elements mainly comes with a set of basis functions in reference coordinates for each applicable AbstractElementGeometry and a generator for the degrees of freedom map of a FESpace

the type steers how the basis functions are transformed from local to global coordinates and how FunctionOperators are evaluated by FEBasisEvaluator.jl

depending on additional continuity properties of the element types more basis function sets are defined:

AbstractH1FiniteElements additionally have evaluations of nonzero basisfunctions on faces/bfaces

AbstractHdivFiniteElements additionally have evaluations of nonzero normalfluxes of basisfunctions on faces/bfaces

List of implemented Finite Elements

The following table lists all curently implemented finite elements. Click on them to find out more details.

H1 finite elements	Hdiv finite elements	Hcurl finite elements	L2 finite elements

H1P1	HDIVRT0		L2P0

H1MINI	HDIVBDM1		L2P1

H1P2	HDIVRT1

H1BR

H1CR

Details

H1-conforming finite elements

GradientRobustMultiPhysics.H1P1 — Type

abstract type H1P1 <: AbstractH1FiniteElement

Continuous piecewise first-order polynomials with arbitrary number of components.

allowed ElementGeometries:

Edge1D (linear polynomials)

Triangle2D (linear polynomials)

Quadrilateral2D (Q1 space)

Tetrahedron3D (linear polynomials)

Hexahedron3D (Q1 space)

GradientRobustMultiPhysics.H1MINI — Type

abstract type H1MINI <: AbstractH1FiniteElement

Mini finite element (continuous piecewise linear + cell bubbles)

allowed element geometries:

Triangle2D (linear polynomials + cubic cell bubble)

Quadrilateral2D (Q1 space + quartic cell bubble)

GradientRobustMultiPhysics.H1P2 — Type

abstract type H1P2 <: AbstractH1FiniteElement

Continuous piecewise second-order polynomials

allowed ElementGeometries:

Edge1D (quadratic polynomials)

Triangle2D (quadratic polynomials)

Quadrilateral2D (Q2 space)

GradientRobustMultiPhysics.H1CR — Type

abstract type H1CR <: AbstractH1FiniteElement

Crouzeix-Raviart element (only continuous at face centers)

allowed ElementGeometries:

Triangle2D (piecewise linear, similar to P1)

Quadrilateral2D (similar to Q1 space)

GradientRobustMultiPhysics.H1BR — Type

abstract type H1BR <: AbstractH1FiniteElementWithCoefficients

vector-valued (ncomponents = edim) Bernardi–Raugel element (first-order polynomials + normal-weighted face bubbles)

allowed ElementGeometries:

Triangle2D (piecewise linear + normale-weighted face bubbles)

Quadrilateral2D (Q1 space + normal-weighted face bubbles)

Hdiv-conforming finite elements

GradientRobustMultiPhysics.HDIVRT0 — Type

abstract type HDIVRT0 <: AbstractHdivFiniteElement

Hdiv-conforming vector-valued (ncomponents = edim) lowest-order Raviart-Thomas space

allowed ElementGeometries:

Triangle2D

Quadrilateral2D

Tetrahedron3D

Hexahedron3D

GradientRobustMultiPhysics.HDIVBDM1 — Type

abstract type HDIVBDM1 <: AbstractHdivFiniteElement

Hdiv-conforming vector-valued (ncomponents = edim) lowest-order Brezzi-Douglas-Marini space

allowed ElementGeometries:

Triangle2D

Quadrilateral2D

GradientRobustMultiPhysics.HDIVRT1 — Type

abstract type HDIVRT1 <: AbstractHdivFiniteElement

Hdiv-conforming vector-valued (ncomponents = edim) Raviart-Thomas space of order 1

allowed ElementGeometries:

Triangle2D

Hcurl-conforming finite elements

L2-conforming finite elements

GradientRobustMultiPhysics.L2P0 — Type

abstract type L2P0 <: AbstractH1FiniteElement

piecewise constant polynomials

allowed on every ElementGeometry

GradientRobustMultiPhysics.L2P1 — Type

abstract type L2P1 <: AbstractH1FiniteElement

discontinuous piecewise first-order polynomials (same as H1P1 but with other dofmap)

allowed ElementGeometries:

Edge1D (linear polynomials)

Triangle2D (linear polynomials)

Quadrilateral2D (Q1 space)