Finite Element Interpolations

Each finite element has its standard interpolator that can be applied to some user-defined DataFunction.

GradientRobustMultiPhysics.interpolate! — Method

function interpolate!(Target::FEVectorBlock,
     AT::Type{<:AbstractAssemblyType},
     source_data::UserData{AbstractDataFunction};
     items = [],
     time = 0,
     verbosity::Int = 0)

Interpolates the given sourcedata into the finite elements space assigned to the Target FEVectorBlock with the specified AbstractAssemblyType (usualy ONCELLS). The optional time argument is only used if the source_data depends on time.

GradientRobustMultiPhysics.interpolate! — Method

function interpolate!(Target::FEVectorBlock,
     source_data::UserData{AbstractDataFunction};
     items = [],
     time = 0,
     verbosity::Int = 0)

Interpolates the given sourcedata into the finite element space assigned to the Target FEVectorBlock. The optional time argument is only used if the sourcedata depends on time.

GradientRobustMultiPhysics.nodevalues! — Function

function nodevalues!(
    Target::AbstractArray{<:Real,2},
    Source::FEVectorBlock,
    operator::Type{<:AbstractFunctionOperator} = Identity;
    regions::Array{Int,1} = [0],
    target_offset::Int = 0,   # start to write into Target after offset
    zero_target::Bool = true, # target vector is zeroed
    continuous::Bool = false)

Evaluates the finite element function with the coefficient vector Source and the specified FunctionOperator at all the nodes of the (specified regions of the) grid and writes the values into Target. Discontinuous (continuous = false) quantities are averaged.