Meshing

Meshes are stored as an ExtendableGrid, see ExtendableGrids.jl for details and constructors. Grid generators for simplex grids can be found e.g. in the external module SimplexGridFactory.jl

Available mesh manipulations

GradientRobustMultiPhysics.barycentric_refine — Method

generates a new ExtendableGrid by barycentric refinement of each cell in the source grid

barcentric refinement is available for these ElementGeometries - Quadrilateral2D (first split into Triangle2D) - Triangle2D

GradientRobustMultiPhysics.split_grid_into — Method

split_grid_into(source_grid::ExtendableGrid{T,K}, targetgeometry::Type{var"#s13"} where var"#s13"<:AbstractElementGeometry) -> ExtendableGrid{_A,_B} where _B where _A

generates a new ExtendableGrid by splitting each cell into subcells of the specified targetgeometry

split rules exist for - Quadrilateral2D into Triangle2D

GradientRobustMultiPhysics.uniform_refine — Method

uniform_refine(source_grid::ExtendableGrid{T,K}) -> ExtendableGrid{_A,_B} where _B where _A

generates a new ExtendableGrid by uniform refinement of each cell in the given grid

uniform refinement rules are available for these AbstractElementGeometries: - Line1D (bisection into two subsegments) - Triangle2D (red refinement into four subtriangles) - Quadrilateral2D (into four subquadrilaterals) - Tetrahedron (into eight subtetrahedrons) - Hexahedron (into eight subhexahedrons)

if multiple geometries are in the mesh uniform refinement will only work if all refinement rules refine faces and edges (in 3D) equally (so no hanging nodes are created)