Meshing

Meshes are stored as an ExtendableGrid, see ExtendableGrids.jl for details and constructors. Grid generators for simplex grids can be found e.g. in the external module SimplexGridFactory.jl

Cells, faces and edges of the mesh are associated to AbstractElementGeometries (defined by ExtendableGrids.jl) that are used to dispatch functionality (local/global transformation, enumeration rules, set of basis functions, volume calculation, refinements etc.). See further below for a list of recognized element geometries.

Available Global Mesh Manipulations

GradientRobustMultiPhysics.RGB_refine — Method

RGB_refine(source_grid::ExtendableGrid{T, K}, facemarkers::Vector{Bool}) -> ExtendableGrid{_A, _B} where {_A, _B}

generates a new ExtendableGrid by red-green-blue mesh refinement of triangular meshes, see e.g.

Carstensen, C. –An Adaptive Mesh-Refining Algorithm Allowing for an H^1 Stable L^2 Projection onto Courant Finite Element Spaces– Constr Approx 20, 549–564 (2004). https://doi.org/10.1007/s00365-003-0550-5

The bool array facemarkers determines which faces should be bisected. Note, that a closuring is performed such that the first face in every triangle with a marked face is also refined.

GradientRobustMultiPhysics.barycentric_refine — Method

generates a new ExtendableGrid by barycentric refinement of each cell in the source grid

barycentric refinement is available for these ElementGeometries

Quadrilateral2D (first split into Triangle2D)

Triangle2D

GradientRobustMultiPhysics.split_grid_into — Method

split_grid_into(source_grid::ExtendableGrid{T, K}, targetgeometry::Type{var"#s63"} where var"#s63"<:AbstractElementGeometry) -> ExtendableGrid{_A, _B} where {_A, _B}

generates a new ExtendableGrid by splitting each cell into subcells of the specified targetgeometry

split rules exist for

Quadrilateral2D into Triangle2D

Hexahedron3D into Tetrahedron3D

GradientRobustMultiPhysics.uniform_refine — Method

generates a new ExtendableGrid by uniform refinement of each cell in the given grid

uniform refinement rules are available for these AbstractElementGeometries:

Line1D (bisection into two subsegments)

Triangle2D (red refinement into four subtriangles)

Quadrilateral2D (into four subquadrilaterals)

Tetrahedron (into eight subtetrahedrons)

Hexahedron (into eight subhexahedrons)

if multiple geometries are in the mesh uniform refinement will only work if all refinement rules refine faces and edges (in 3D) equally (so no hanging nodes are created)

GradientRobustMultiPhysics.reference_domain — Function

    reference_domain(EG::Type{<:AbstractElementGeometry}, T::Type{<:Real} = Float64; scale = [1,1,1], shift = [0,0,0]) -> ExtendableGrid{T,Int32}

Generates an ExtendableGrid{T,Int32} for the reference domain of the specified Element Geometry. With scale and shift the coordinates can be manipulated.

Recognized Geometries and Reference Domains

The following list contains all recognized subtypes of ExtendableGrids.AbstractElementGeometries and their reference domains. Moreover, each geometry has a number of rules that define the geometries and the local enumeration of its nodes, faces and edges, see source code of shape_specs.jl for details

Edge1D <: AbstractElementGeometry1D

[1]-----[2]               [1] = [0]
                          [2] = [1]

Triangle2D

[3]                 
 | \   
 |   \                    [1] = [0,0]
 |     \                  [2] = [1,0]
 |       \                [3] = [0,1]
 |         \ 
[1]--------[2]

Parallelogram2D <: Quadrilateral2D

[4]--------[3]               
 |          |             [1] = [0,0]
 |          |             [2] = [1,0]
 |          |             [3] = [1,1]
 |          |             [4] = [0,1]
[1]--------[2]

Note: most finite elements only work as intended on Parallelogram2D
      since the local<>global map stays affine in this case

Tetrahedron3D

[4]                 
 |\\   
 | \ \                    [1] = [0,0,0]
 |  \  \                  [2] = [1,0,0]
 |   \   \                [3] = [0,1,0]
 | _-[3]-_ \              [4] = [0,0,1]
[1]--------[2]

Parallelepiped3D <: Hexahedron3D

    [8]--------[7]        [1] = [0,0,0]
   / |        / |         [2] = [1,0,0]
[5]--------[6]  |         [3] = [1,1,0]
 |   |      |   |         [4] = [0,1,0]
 |   |      |   |         [5] = [0,0,1]
 |  [4]-----|--[3]        [6] = [1,0,1]
 | /        | /           [7] = [1,1,1]
[1]--------[2]            [8] = [0,1,1]

Note: most finite elements only work as intended on Parallelepiped3D
      since the local<>global map stays affine in this case