PDE Operators

Purpose

The PDE consists of PDEOperators characterising some feature of the model (like friction, convection, exterior forces etc.), they describe the continuous weak form of the PDE.

GradientRobustMultiPhysics.PDEOperator — Type

mutable struct PDEOperator{T<:Real, APT<:AssemblyPatternType, AT<:AbstractAssemblyType} <: AbstractPDEOperator

common structures for all finite element operators that are assembled with GradientRobustMultiPhysics; better look at the AssemblyPatternType and the constructors

The following table lists all available operators and physics-motivated constructors for them. Click on them or scroll down to find out more details.

Main constructors	Special constructors	Mathematically

AbstractBilinearForm		$(\mathrm{A}(\mathrm{FO}_1(u)),\mathrm{FO}_2(v))$ or $(\mathrm{FO}_1(u),\mathrm{A}(\mathrm{FO}_2(v)))$

LaplaceOperator	$(\kappa \nabla u,\nabla v)$

ReactionOperator	$(\alpha u, v)$

LagrangeMultiplier	$(\mathrm{FO}_1(u), v)$ (automatically assembles 2nd transposed block)

ConvectionOperator	$(\beta \cdot \nabla u, v)$ (beta is function)

HookStiffnessOperator2D	$(\mathbb{C} \epsilon(u),\epsilon(v))$ (also 1D or 3D variants exist)

AbstractTrilinearForm		$(\mathrm{A}(\mathrm{FO}_1(a),\mathrm{FO}_2(u)),\mathrm{FO}_3(v))$

ConvectionOperator	$((a \cdot \nabla) u, v)$ (a is registered unknown)

ConvectionRotationFormOperator	$((a \times \nabla) u,v)$ (a is registered unknown, only 2D for now)

GenerateNonlinearForm		$(\mathrm{NA}(\mathrm{FO}_1(u),...,\mathrm{FO}_{N-1}(u)),\mathrm{FO}_N(v))$

RhsOperator		$(f \cdot \mathrm{FO}(v))$

Legend: $\mathrm{FO}$ are placeholders for Function Operators, and $\mathrm{A}$ stands for a (linear) Action (that only expects the operator value of the finite element function as an input) and $\mathrm{NA}$ stands for a (nonlinear) Action (see GenerateNonlinearForm for details).

Assembly Type

Many PDE operators need a specification of that decides to which parts of the mesh the PDEOperator is associated (e.g. cells, faces, bfaces, edges), this is prescribed via the AssemblyType. The following assembly types are available. Additional to define where PDEOperators live and assemble, they can be also used as an argument for interpolation!.

AssemblyType	Description

AT_NODES	interpolate at vertices of the mesh (only for H1-conforming FEM)

ON_CELLS	assemble/interpolate over the cells of the mesh

ON_FACES	assemble/interpolate over all faces of the mesh

ON_IFACES	assemble/interpolate over the interior faces of the mesh

ON_BFACES	assemble/interpolate over the boundary faces of the mesh

ON_EDGES (*)	assemble/interpolate over all edges of the mesh (in 3D)

ON_BEDGES (*)	assemble/interpolate over the boundary edges of the mesh (in 3D)

Note

(*) = only reasonable in 3D and still experimental, might have some issues

GradientRobustMultiPhysics.AT_NODES — Type

abstract type AT_NODES <: AbstractAssemblyType

causes interpolation at vertices of the grid (only for H1-conforming interpolations)

GradientRobustMultiPhysics.ON_BEDGES — Type

abstract type ON_BEDGES <: AbstractAssemblyType

causes assembly/interpolation on boundary edges of the grid (only in 3D)

GradientRobustMultiPhysics.ON_BFACES — Type

abstract type ON_BFACES <: AbstractAssemblyType

causes assembly/interpolation on boundary faces of the grid

GradientRobustMultiPhysics.ON_CELLS — Type

abstract type ON_CELLS <: AbstractAssemblyType

causes assembly/interpolation on cells of the grid

GradientRobustMultiPhysics.ON_EDGES — Type

abstract type ON_EDGES <: AbstractAssemblyType

causes assembly/interpolation on edges of the grid (only in 3D)

GradientRobustMultiPhysics.ON_FACES — Type

abstract type ON_FACES <: AbstractAssemblyType

causes assembly/interpolation on faces of the grid

GradientRobustMultiPhysics.ON_IFACES — Type

abstract type ON_IFACES <: ON_FACES

causes assembly/interpolation on interior faces of the grid

Special Linear Operators

Below you find the special constructors of available common linear, bilinear and trilinear forms.

GradientRobustMultiPhysics.LaplaceOperator — Function

LaplaceOperator() -> GradientRobustMultiPhysics.PDEOperator{Float64, SymmetricBilinearForm, ON_CELLS}
LaplaceOperator(diffusion::Any; name, AT, gradient_operator, regions, store) -> Union{Nothing, GradientRobustMultiPhysics.PDEOperator{Float64, SymmetricBilinearForm, ON_CELLS}}

constructor for a bilinearform that describes a(u,v) = (kappa * nabla u, nabla v) where kappa is some constant diffusion coefficient

GradientRobustMultiPhysics.ReactionOperator — Function

ReactionOperator() -> GradientRobustMultiPhysics.PDEOperator{Float64, SymmetricBilinearForm, ON_CELLS}
ReactionOperator(coefficient::Any; name, AT, identity_operator, regions, store) -> Union{Nothing, GradientRobustMultiPhysics.PDEOperator{Float64, SymmetricBilinearForm, ON_CELLS}}

constructor for a bilinearform that describes a(u,v) = (A(u),v) or (u,A(v)) with some user-specified action A

GradientRobustMultiPhysics.ConvectionOperator — Function

function ConvectionOperator(
    a_from::Int, 
    beta_operator::Type{<:AbstractFunctionOperator},
    xdim::Int,
    ncomponents::Int;
    name = "auto",
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    fixed_argument::Int = 1,
    factor = 1,
    ansatzfunction_operator::Type{<:AbstractFunctionOperator} = Gradient,
    testfunction_operator::Type{<:AbstractFunctionOperator} = Identity,
    regions::Array{Int,1} = [0],
    auto_newton::Bool = false,
    quadorder = 0)

constructs a trilinearform for a convection term of the form c(a,u,v) = (betaoperator(a)*grad(u),v) where afrom is the id of some unknown of the PDEDescription. xdim is the space dimension (= number of components of betaoperato(a)) and ncomponents is the number of components of u. With fixedargument = 2 a and u can switch their places, i.e. c(u,a,v) = (betaoperator(u)*grad(a),v), With autonewton = true a Newton scheme for a(u,v) = (u*grad(u),v) is automatically derived (and fixed_argument is ignored).

ConvectionOperator(beta::UserData{AbstractDataFunction, FType, NFType, ndim} where {FType<:Function, NFType<:Function, ndim}, ncomponents::Int64; name, store, AT, ansatzfunction_operator, testfunction_operator, regions) -> GradientRobustMultiPhysics.PDEOperator{Float64, BilinearForm, ON_CELLS}

constructor for a bilinearform that describes a(u,v) = (beta*grad(u),v) with some user-specified DataFunction beta that writes into an result array of length ncomponents

GradientRobustMultiPhysics.ConvectionRotationFormOperator — Function

ConvectionRotationFormOperator(beta::Int64, beta_operator::Type{var"#s181"} where var"#s181"<:??, xdim::Int64, ncomponents::Int64; name, AT, factor, ansatzfunction_operator, testfunction_operator, regions) -> Union{Nothing, GradientRobustMultiPhysics.PDEOperator{Float64, TrilinearForm, ON_CELLS}}

constructor for a trilinearform that describes a(u,v) = (beta x curl(u),v) where beta is the id of some unknown vector field of the PDEDescription, u and v are also vector-fields and x is the cross product (so far this is only implemented in 2D)

GradientRobustMultiPhysics.HookStiffnessOperator1D — Function

HookStiffnessOperator1D(mu::Any; name, regions, gradient_operator, store) -> GradientRobustMultiPhysics.PDEOperator{Float64, BilinearForm, ON_CELLS}

constructor for a bilinearform that describes a(u,v) = (C grad(u), grad(v)) where C is the 1D stiffness tensor C grad(u) = mu grad(u)

GradientRobustMultiPhysics.HookStiffnessOperator2D — Function

HookStiffnessOperator2D(mu::Any, lambda::Any; name, AT, regions, gradient_operator, store) -> GradientRobustMultiPhysics.PDEOperator{Float64, BilinearForm, ON_CELLS}

constructor for a bilinearform that describes a(u,v) = (C eps(u), eps(v)) where C is the 3D stiffness tensor for isotropic media in Voigt notation, i.e. C eps(u) = 2 mu eps(u) + lambda tr(eps(u)) for Lame parameters mu and lambda

In Voigt notation C is a 3 x 3 matrix
C = [c11,c12,  0
     c12,c11,  0
       0,  0,c33]

where c33 = shear_modulus, c12 = lambda and c11 = 2*c33 + c12

GradientRobustMultiPhysics.HookStiffnessOperator3D — Function

HookStiffnessOperator3D(mu::Any, lambda::Any; name, AT, regions, gradient_operator, store) -> GradientRobustMultiPhysics.PDEOperator{Float64, BilinearForm, ON_CELLS}

constructor for a bilinearform that describes a(u,v) = (C eps(u), eps(v)) where C is the 3D stiffness tensor for isotropic media in Voigt notation, i.e. C eps(u) = 2 mu eps(u) + lambda tr(eps(u)) for Lame parameters mu and lambda

In Voigt notation C is a 6 x 6 matrix
C = [c11,c12,c12,  0,  0,  0
     c12,c11,c12,  0,  0,  0
     c12,c12,c11,  0,  0,  0
       0,  0,  0,c44,  0,  0
       0,  0,  0,  0,c44,  0
       0,  0,  0,  0,  0,c44]   

where c44 = shear_modulus, c12 = lambda and c11 = 2*c44 + c12

GradientRobustMultiPhysics.RhsOperator — Function

RhsOperator(operator::Type{var"#s205"} where var"#s205"<:??, action::AbstractAction; name, AT, regions, factor, store) -> GradientRobustMultiPhysics.PDEOperator{Float64, LinearForm, ON_CELLS}

generates a linearform from an action

RhsOperator(operator::Type{var"#s179"} where var"#s179"<:??, regions::Vector{Int64}, data::UserData{var"#s127", FType, NFType, ndim} where {var"#s127"<:AbstractDataFunction, FType<:Function, NFType<:Function, ndim}; name, AT, factor, store) -> GradientRobustMultiPhysics.PDEOperator{Float64, LinearForm, ON_CELLS}

generates a linearform from a given UserData{<:DataFunction} (whose result dimension has to be 1)

Custom Linear Operators

It is possible to define custom bilineraforms and trilinearforms by specifiyng Function Operators and (in case of bilinearform optionally) an Action.

GradientRobustMultiPhysics.AbstractBilinearForm — Function

function AbstractBilinearForm(
    operators::Array{AbstractFunctionOperator,1},
    action::AbstractAction = NoAction();
    name = "auto",
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    apply_action_to = 1,
    regions::Array{Int,1} = [0],
    transposed_assembly::Bool = false,
    store::Bool = false)

abstract bilinearform constructor that assembles

b(u,v) = intregions action(operator1(u)) * operator2(v) if applyaction_to = 1

b(u,v) = intregions operator1(u) * action(operator2(v)) if applyaction_to = 2

The optional arguments AT and regions specifies on which grid item the operator lives/assembles, while store toggles the separate storage for the operator (which is advisable if it is not alone i an otherweise nonlinear block of a PDEDescription).

GradientRobustMultiPhysics.AbstractTrilinearForm — Function

function AbstractTrilinearForm(
    operators::Array{AbstractFunctionOperator,1},
    a_from::Int,
    a_to::Int,
    action::AbstractAction;
    name = "auto",
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    regions::Array{Int,1} = [0],
    transposed_assembly::Bool = false)

abstract trilinearform constructor that assembles

c(a,u,v) = int_regions action(operator1(a),operator2(u)) * operator3(v)

where u and are the ansatz and test function coressponding to the PDE coordinates and a is an additional unknown of the PDE. The argument a can be moved to the other positions with ato and gets it data from unknown afrom of the full PDEdescription.

The optional arguments AT and regions specifies on which grid item the operator lives/assembles,

Also note that this operator is always marked as nonlinear by the Solver configuration.

Examples

Below some examples for operators are given:

# Example 1 : div-div bilinearform with a factor λ (e.g. for divergence-penalisation)
operator = AbstractBilinearForm([Divergence,Divergence]; name = "λ (div(u),div(v))", factor = λ)

# Example 2 : Gradient jump stabilisation with an item-dependent action and a factor s (e.g. for convection stabilisation)
xFaceVolumes::Array{Float64,1} = xgrid[FaceVolumes]
function stabilisation_kernel(result, input, item)
    result .= input 
    result .*= xFaceVolumes[item]^2
end
action = Action(Float64,stabilisation_kernel, [2,2]; dependencies = "I", quadorder = 0 )
operator = AbstractBilinearForm([Jump(Gradient), Jump(Gradient)], action; AT = ON_IFACES, name = "s |F|^2 [∇(u)]⋅[∇(v)]", factor = s)

Lagrange Multipliers

There is a special bilinearform intended to use for the assembly of Lagrange multipliers that automatically copies itself to the transposed block of the PDEdescription.

GradientRobustMultiPhysics.LagrangeMultiplier — Function

LagrangeMultiplier(operator::Type{var"#s205"} where var"#s205"<:??; name, AT, action, regions, store, factor) -> GradientRobustMultiPhysics.PDEOperator{Float64, BilinearForm, ON_CELLS}

constructor for a bilinearform that describes a(u,v) = (A(operator(u)), id(v)) and assembles a second transposed block at the block of the transposed PDE coordinates. It is intended to use to render one unknown of the PDE the Lagrange multiplier for another unknown by putting this operator on the coressponding subdiagonal block of the PDE description.

Example: LagrangeMultiplier(Divergence) is used to render the pressure the LagrangeMultiplier for the velocity divergence constraint in the Stokes prototype.

Nonlinear Operators

Nonlinear Operators can be setup in two ways. The manual way requires the user to define an action with a nonlinear action kernel (see Action Kernel) that specifies the linearisation of the nonlinearity. There is also an automatic way where the user specifies only a function (where the input can be used nonlinearly) which is then automatically differentiated to generate the linearised action kernel, see below for details.

GradientRobustMultiPhysics.GenerateNonlinearForm — Function

function GenerateNonlinearForm(
    name::String,
    operator1::Array{DataType,1},
    coeff_from::Array{Int,1},
    operator2::Type{<:AbstractFunctionOperator},
    action_kernel::Function,
    argsizes::Array{Int,1},
    dim::Int;
    AT::Type{<:AbstractAssemblyType} = ON_CELLS,
    ADnewton::Bool = false,
    action_kernel_rhs = nothing,
    regions = [0])

generates an abstract nonlinearform operator G. The array coeff_from stores the ids of the unknowns that should be used to evaluate the operators. The array argsizes is a vector with two entries where the first one is the length of the expected result vector and the second one is the length of the input vector.

If ADnewton == true, the specified actionkernel is automatically differentiated to assemble the Jacobian DG and setup a Newton iteration. The actionkernel has to be a function of the interface

function name(result,input)

where input is a vector of the operators of the solution and result is what then is multiplied with operator2 of the testfunction. Given some operator G(u), the Newton iteration reads DG u_next = DG u - G(u) which is added to the rest of the (linear) operators in the PDEDescription.

If ADnewton == false, the user is epected to prescribe a linearisation of the nonlinear operator. In this case the action_kernel has to satisfy the interface

function name(result, input_current, input_ansatz)

where inputcurrent is a vector of the operators of the solution and inputansatz is a vecor with the operators evaluated at one of the basis functions. If necessary, also a right-hand side action in the same format can be prescribed in actionkernelrhs.

Note: this is a highly experimental feature at the moment and will possibly only work when all operators are associated with the same unknown.

can only be applied in PDE LHS

Other Operators

There are some more operators that do not fit into the structures above. Also, in the future, the goal is to open up the operator level for exterior code to setup operators that are assembled elsewhere.

GradientRobustMultiPhysics.FVConvectionDiffusionOperator — Type

FVConvectionDiffusionOperator(beta_from::Int64; diffusion) -> FVConvectionDiffusionOperator

finite-volume convection diffusion operator (for cell-wise P0 rho)

div(diffusion * grad(rho) + beta rho)

For diffusion = 0, the upwind divergence: div_upw(beta*rho) is generated For diffusion > 0, TODO

GradientRobustMultiPhysics.DiagonalOperator — Type

DiagonalOperator() -> DiagonalOperator
DiagonalOperator(value::Real) -> DiagonalOperator
DiagonalOperator(value::Real, onlynz::Bool; regions) -> DiagonalOperator

puts value on the diagonal entries of the cell dofs within given regions

if onlyz == true only values that are zero are changed

can only be applied in PDE LHS

GradientRobustMultiPhysics.CopyOperator — Type

CopyOperator(copy_from::Any, factor::Any) -> CopyOperator

copies entries from TargetVector to rhs block

can only be applied in PDE RHS