PDE Solvers

If solve! is applied to a PDEDescription and a FEVector (that specifies the ansatz spaces for the unknowns) an investigation of the PDEDescription is performed that decides if the problem is nonlinear (and has to be solved by a fixed-point algorithm) or if it can be solved directly in one step. Additionally the user can manually trigger subiterations that splits the fixed-point algorithm into substeps where only subsets of the PDE equations are solved together.

Also, there is a preliminary time-dependent solver that can be setup in a similar manner and then performs the subiterations once in eachtimestep. As a TimeIntegrationRule so far only BackwardEuler is implemented.

GradientRobustMultiPhysics.TimeControlSolver — Type

function TimeControlSolver(
    PDE::PDEDescription,
    Solution::FEVector,    # contains initial values and stores solution of advance method below
    TIR::Type{<:AbstractTimeIntegrationRule} = BackwardEuler;
    timedependent_equations = [],
    subiterations = "auto",
    start_time::Real = 0,
    verbosity::Int = 0,
    dt_testfunction_operator = [],
    dt_action = [],
    nonlinear_dt::Bool = false,
    dirichlet_penalty = 1e60)

Creates a time-dependent solver that can be advanced in time with advance!. The FEVector Solution stores the initial state but also the solutions of each timestep. The field timedependent_equations contains the equation numbers (=rows in PDEDescription) that are time-dependent. The field subiterations specifies subsets of equations that are solved together in the given order; "auto" tries to solve the whole system at once.

GradientRobustMultiPhysics.advance! — Function

advance!(TCS::TimeControlSolver) -> Array{Float64,2}
advance!(TCS::TimeControlSolver, timestep::Real) -> Array{Float64,2}

Advances a TimeControlSolver in time with the given timestep.

GradientRobustMultiPhysics.solve! — Method

solve!(Target::FEVector, PDE::PDEDescription; subiterations, dirichlet_penalty, time, maxResidual, maxIterations, linsolver, maxlureuse, AndersonIterations, verbosity)

Solves a given PDE (provided as a PDEDescription) and writes the solution into the FEVector Target. The ansatz spaces are taken from the of this vector. The field subiterations specifies subsets of equations that are solved together in the given order; "auto" tries to solve the whole system at once, but also might resort to a fixed-point iteration if a nonlinearity is detected.