PDE Solvers

Fixed-Time Solvers

If solve! is applied to a PDEDescription and a FEVector (that specifies the ansatz spaces for the unknowns) an investigation of the PDEDescription is performed that decides if the problem is nonlinear (and has to be solved by a fixed-point algorithm) or if it can be solved directly in one step. Additionally the user can manually trigger subiterations that splits the fixed-point algorithm into substeps where only subsets of the PDE equations are solved together.

GradientRobustMultiPhysics.solve! — Function

function solve!(
    Target::FEVector,
    PDE::PDEDescription;
    subiterations = "auto",
    dirichlet_penalty::Real = 1e60,
    time::Real = 0,
    maxResidual::Real = 1e-12,
    maxIterations::Int = 10,
    linsolver = DirectUMFPACK,
    maxlureuse = [1],
    AndersonIterations = 0, #  0 = Picard iteration, >0 Anderson iteration (experimental feature)
    verbosity::Int = 0)

Solves a given PDE (provided as a PDEDescription) and writes the solution into the FEVector Target. The ansatz spaces are taken from the of this vector.

Further optional arguments:

subiterations : specifies subsets of equations that are solved together in the given order; "auto" tries to solve the whole system at once

dirichlet_penalty : Dirichlet data is enforced by penalties on the diagonal of the matrix.

time : If time-dependent data is involved, the time can be fixed to some value here.

linsolver : specifies the linear solver that is used to solver the linear system of equation in each fixpoint iteration

maxlureuse : specifies after how much iterations the lu decomposition should be recomputed (-1 = only once if e.g. the matrix stays the same for each fixpoint iteration)

Depending on the subiterations and detected/configured nonlinearties the whole system is either solved directly in one step or via a fixed-point iteration.

Time-Dependent Solvers

The structure TimeControlSolver can be used to setup a time-dependent solver that can be configured in a similar manner as the time-independent ones (subiterations, nonlinear iterations, linear solvers). As a TimeIntegrationRule so far only BackwardEuler is implemented. It detects automatically when the timestep changes and when the time derivative has to be reimplemented. Also non-standard and parameter-dependent time derivatives are possible.

Note that, the time-derivative is added by the TimeControlSolver and is in general not part of the PDEDescription (this is debatable and might change in the future).

GradientRobustMultiPhysics.TimeControlSolver — Type

function TimeControlSolver(
    PDE::PDEDescription,
    Solution::FEVector,    # contains initial values and stores solution of advance method below
    TIR::Type{<:AbstractTimeIntegrationRule} = BackwardEuler;
    timedependent_equations = [],
    subiterations = "auto",
    start_time::Real = 0,
    verbosity::Int = 0,
    dt_testfunction_operator = [],
    dt_action = [],
    nonlinear_dt::Bool = false,
    dirichlet_penalty = 1e60)

Creates a time-dependent solver that can be advanced in time with advance!. The FEVector Solution stores the initial state but also the solutions of each timestep that are computed via any of the advance! methods. The argument TIR carries the time integration rule (currently there is only BackwardEuler).

Further optional arguments (that are not listed in the description of solve!):

timedependent_equations : contains the equation numbers (=rows in PDEDescription) that are time-dependent and should get a time derivative (currently only time derivatives of order 1)

start_time : initial time

dttestfunction_operator : operator applied to testfunctions in time derivative

dt_action : additional actions that are applied to the ansatz function in the time derivative (to include parameters etc.)

GradientRobustMultiPhysics.advance! — Function

function TimeControlSolver(
    advance!(TCS::TimeControlSolver, timestep::Real = 1e-1)

Advances a TimeControlSolver one step in time with the given timestep.

Moreover there are two functions that advance the TimeControlSolver automatically until a given final time (advance_until_time!) is reached or until stationarity is reached (advance_until_stationarity!). At the moment only uniform timestepping is used, but adaptive time stepping might be implemented at some point.

GradientRobustMultiPhysics.advance_until_time! — Function

advance_until_time!(TCS::TimeControlSolver, timestep, finaltime; finaltime_tolerance = 1e-15, do_after_each_timestep = nothing)

Advances a TimeControlSolver in time with the given (initial) timestep until the specified finaltime is reached (up to the specified tolerance). The function doaftertimestep is called after each timestep and can be used to print/save data (and maybe timestep control in future).

GradientRobustMultiPhysics.advance_until_stationarity! — Function

advance_until_stationarity!(TCS::TimeControlSolver, timestep; stationarity_threshold = 1e-11, maxTimeSteps = 100, do_after_each_timestep = nothing)

Advances a TimeControlSolver in time with the given (initial) timestep until stationarity is detected (change of variables below threshold) or a maximal number of time steps is exceeded. The function doaftertimestep is called after each timestep and can be used to print/save data (and maybe timestep control in future).