Quadrature

Quadrature rules consist of points (coordinates of evauation points with respect to reference geometry) and weights. There are constructors for several AbstractElementGeometries (from ExtendableGrids) and different order (some have generic formulas for abitrary order), see below for a detailed list.

GradientRobustMultiPhysics.QuadratureRule — Type

abstract type QuadratureRule{T<:Real, ET<:AbstractElementGeometry}

A struct that contains the name of the quadrature rule, the reference points and the weights for the parameter-determined element geometry.

GradientRobustMultiPhysics.QuadratureRule — Method

function QuadratureRule{T,ET}(order::Int) where {T<:Real, ET <: AbstractElementGeometry0D}

Constructs 0D quadrature rule of specified order (always point evaluation).

GradientRobustMultiPhysics.QuadratureRule — Method

function QuadratureRule{T,ET}(order::Int) where {T<:Real, ET <: AbstractElementGeometry1D}

Constructs 1D quadrature rule of specified order.

GradientRobustMultiPhysics.QuadratureRule — Method

function QuadratureRule{T,ET}(order::Int) where {T<:Real, ET <: Parallelepiped3D}

Constructs quadrature rule on Parallelepiped3D of specified order.

GradientRobustMultiPhysics.QuadratureRule — Method

function QuadratureRule{T,ET}(order::Int) where {T<:Real, ET <: Parallelogram2D}

Constructs quadrature rule on Parallelogram2D of specified order.

GradientRobustMultiPhysics.QuadratureRule — Method

function QuadratureRule{T,ET}(order::Int) where {T<:Real, ET <: Tetrahedron3D}

Constructs quadrature rule on Tetrahedron3D of specified order.

GradientRobustMultiPhysics.QuadratureRule — Method

function QuadratureRule{T,ET}(order::Int) where {T<:Real, ET <: Triangle2D}

Constructs quadrature rule on Triangle2D of specified order.

Base.eltype — Method

eltype(_::QuadratureRule{T<:Real, ET<:AbstractElementGeometry}) -> Vector{_A} where _A

Custom eltype function for QuadratureRule{T,ET}.

Base.show — Method

show(io::IO, Q::QuadratureRule)

Custom show function for QuadratureRule{T,ET} that prints some information.

GradientRobustMultiPhysics.integrate! — Method

integrate!(integral4items::AbstractArray, grid::ExtendableGrid, AT::Type{var"#s114"} where var"#s114"<:AbstractAssemblyType, integrand::UserData{var"#s113", ndim} where {var"#s113"<:AbstractDataFunction, ndim}; index_offset, time, items, force_quadrature_rule)

Integration that writes result on every item into integral4items.

GradientRobustMultiPhysics.integrate — Method

integrate(grid::ExtendableGrid, AT::Type{var"#s181"} where var"#s181"<:AbstractAssemblyType, integrand!::UserData{var"#s180", ndim} where {var"#s180"<:AbstractDataFunction, ndim}, resultdim::Int64; items, force_quadrature_rule) -> Union{Float64, Vector{Float64}}

Integration that returns total integral.