User Data and Actions

There is a variety of different user data, like scalar- and vector-valued constants, time-dependent data, region-dependent data or plain functions that depend on the the space coordinates. Also dependency on the item number of the reference coordinates of the quadrature point in the quadrature item are sometimes desireable. To allow for flexible user-specified data, all functions have to be negotiated by the UserData interface that fixes the order and number of the arguments in the interface via a user-given substring of "XTRIL" where each character stands for a dependency. The following table explains the meaning of each character.

Character	Explanation

X	depends on (vector-valued) space coordinates

T	depends on time coordinate

R	depends on region number

I	depends on item number

L	depends on local coordinates in reference geometry of item

Also note that all functions are expected to write their result into the first argument.

Data Function

The simplest form of user data is called DataFunction which allows additional dependencies on space or time coordinates. The following tables lists all allowed substrings of "XTRIL" and the expected interface of the function provided by the user.

dependency string	Expected interface

""	function f!(result) ... end (constant data)

"X"	function f!(result,x) ... end (space-dependent data)

"T"	function f!(result,t) ... end (time-dependent constant-in-space data)

"XT"	function f!(result,x,t) ... end (space and time-dependent data)

DataFunctions can be used to define boundary data, right-hand side functions and can be interpolated by the finite element standard interpolations.

GradientRobustMultiPhysics.DataFunction — Function

function DataFunction(
    f::Function,                    # user function with interface f(result, _other dependencies_)
    dimensions::Array{Int,1};       # [length(result), length(x)]
    name = "user data function",
    dependencies::String = "",      # substring of "XT" encoding other dependencies in f interface
    quadorder::Int = 0)             # quadrature order added to operator's quadorder that evalute f

Provides a negotation interface for some user-defined function that can be used in integrate! and boundary or right-hand side data assignments. The function f has to obey the interface

f(result, [X, T])

where the parameters X (= space coordinates) and T ( = time) are optional. Which of them are used has to be specified in the String dependencies. The array dimensions specifies the expected length of result and input and quadorder determines the additional quadrature order to be used if this function is involved in some quadrature-requireing procedure.

function DataFunction(c::Array{<:Real,1}; name = "constant user data", quadorder::Int = 0)

Directly generates a DataFunction from a given array c, i.e. a DataFunction that is constant and has no dependencies on x or t.

Extended Data Function

There are also ExtendedDataFunction that allow the additional dependencies R (region), I (item number) and L (local coordinates). The dependencies are stated via a string in the constructor that should be a substring of "XTRIL". However, extended data functions cannot be used everywhere.

GradientRobustMultiPhysics.ExtendedDataFunction — Function

function ExtendedDataFunction(
    f::Function,                    # user function with interface f(result, _other dependencies_)
    dimensions::Array{Int,1};       # [length(result), length(x)]
    name = "user data function",
    dependencies::String = "",      # substring of "XTRIL" encoding other dependencies in f interface
    quadorder::Int = 0)             # quadrature order added to operator's quadorder that evalute f

Provides a negotation interface for some data function with extended dependencies (region number, item number and local coordinates) that can be used in integrate!. The function f has to obey the interface

f(result, [X, T, R, I, L])

where the parameters X (= space coordinates) and T ( = time) are optional. Which of them are used has to be specified in the String dependencies. The array dimensions specifies the expected length of result and X (if X-depdendent, otherwise will be ignored) and quadorder determines the additional quadrature order to be used if this function is involved in some quadrature-requireing procedure.

Action Kernel

Another form of user data are action kernels that are used to define an AbstractAction. Actions modify (usually a subset of) arguments of Assembly Patterns and so allow parameter-dependent assemblies. Also, a trilinear form always needs an action that holds instructions how to prepare the first two arguments such that it can be evaluated with the testfunction operator. To use them, the user defines some kernel function for the action that has the interface. Kernel functions also allow the full range of dependencies, hence any substring of "XTRIL".

However, between the result argument and the further dependencies they get an input argument which (during assembly) carries the operator-evaluations of the arguments that go into the action. Hence the usual interface of a action kernel function looks like this:

function action_kernel!(result,input,[X,T,R,I,L])
    # result = modified input, e.g.
    # multiplication with some parameter that can depend on
    # X = space coordinates
    # T = time
    # R = region number
    # I = item number (cell, face or edge number depending on assembly type)
    # L = local coordinates on item reference domain
end

During assembly, input (in general) takes the role of all non-testfunction arguments and the result vector will be the one that is multiplied with the testfunctions. Additionally, the kernel function can depend on X, T, R, I and L as specified above. Once again note, that time-dependency of the kernel function is inherited to the action and later to the whole PDEOperator and so triggers reassembly of the associated PDEoperator in each time step.

GradientRobustMultiPhysics.ActionKernel — Function

function ActionKernel(
    f::Function,                    # user function with interface f(result, _other dependencies_)
    dimensions::Array{Int,1};       # [length(result), length(input)]
    name = "user action kernel",
    dependencies::String = "",      # substring of "XTRIL" encoding other dependencies in f interface
    quadorder::Int = 0)             # quadrature order added to actions/operators that evaluate this action kernel

Provides a negotation interface for some function that can be used in the Action constructor to define a user-defined operator action. The function has to obey the interface

f(result, input, [X, T, R, I, L])

where the parameters X (= space coordinates), T ( = time), R (= region number), I (= item number), L (= local coordinates) are optional. Which of them are used has to be specified in the String dependencies.

The input vector usually provides the FunctionOperator evaluations of (a subset of) the ansatz arguments of the assembly pattern where the action is used. The array dimensions specifies the expected length of result and input and quadorder determines the additional quadrature order to be used if this function (or its derived action) is involved in an assembly process.

Action

Actions are used by abstract user-defined PDEOperators and consist of an action kernel plus some additional infrastructure. To generate an action from an action kernel or directly from a function works via the following constructors.

GradientRobustMultiPhysics.Action — Type

function Action(
    T::Type{<:Real},
    kernel::UserData{<:AbstractActionKernel};
    name::String = "user action")

Creates an Action from a given specified action kernel that then can be used in an assembly pattern. T specifies the number format that should match the number format of the used quadrature rules and grid coordinates in the mesh (usually T).

function Action{T <: Real}(
    kernel_function::Function,
    dimensions::Array{Int,1};
    name = "user action",
    dependencies = "",
    quadorder = 0)

Creates an Action directly from a kernel function (plus additional information to complement the action kernel) that then can be used in an assembly pattern. T specifies the number format that should match the number format of the used quadrature rules and grid coordinates in the mesh (default if omitted: T = Float64).

NLAction Kernel

For the manual linearisation (=without automatic differentiation) of Nonlinear Operators assembly pattern, the user can also insert nonlinear action kernels that have a second input argument for the operator-evaluations of the current solution. But since this feature is still experimental, it will not explained in more detail, yet.

GradientRobustMultiPhysics.NLActionKernel — Function

function NLActionKernel(
    f::Function,
    dimensions::Array{Int,1};
    name = "nonlinear user action kernel",
    dependencies::String = "",
    quadorder::Int = 0)

Provides a negotation interface for some function to be used as a nonlinear action kernel that can be used in the NonlinearOperator constructor without automatic differentiation. The function f has to obey the interface

f(result, input_current, input_ansatz)

and can be seen as a linearisation of a nonlinearity that can depend on the operator evaluation of the current iterate (inputcurrent) and, as usual, the operator evaluations of the ansatz function (inputansatz).

No further dependencies are allowed currently. Note, that this is a work-in-progress feature.