User Data and Action Kernels

There is a variety of different user data, like scalar- and vector-valued constants, time-dependent data, region-dependent data or plain functions that depend on the the space coordinates. To allow for flexible user-specified data, all functions have to be negotiated by the UserData interface that fixes the order and number of dependencies.

Data Functions

The simplest form of user data is called DataFunction and satisfies the interface

function data_function!(result,x::Array{<:Real,1},t::Real)
    # definition of result, possibly dependent on
    # X = space coordinates
    # T = time
end

data_function! (generic function with 1 method)

where the result vector carries the result of the function. Additionally, it can (but not has to) depend on the further inputs X (space coordinates) or T (time) DataFunctions can be used to define boundary data, right-hand side functions and can be interpolated by the finite element standard interpolations.

There are also ExtendedDataFunction that can additionally depend on R (region), I (item number) and L (local coordinates). The dependencies are stated via a string in the constructor that should be a substring of "XTRIL". However, extended data functions cannot be used everywhere.

GradientRobustMultiPhysics.DataFunction — Function

function DataFunction(
    f::Function,
    dimensions::Array{Int,1};
    name = "user data function",
    dependencies::String = "",
    quadorder::Int = 0)

Provides a negotation interface for some user-defined function that can be used in integrate! and boundary or right-hand side data assignments. The function f has to obey the interface

f(result, [X, T])

where the parameters X (= space coordinates) and T ( = time) are optional. Which of them are used has to be specified in the String dependencies. The string "X" for example specifies that the interface is only space-dependent, i.e

f(result, X)

The array dimensions specifies the expected length of result and input and quadorder determines the additional quadrature order to be used if this function is involved in some quadrature-requireing procedure.

GradientRobustMultiPhysics.ExtendedDataFunction — Function

function ExtendedDataFunction(
    f::Function,
    dimensions::Array{Int,1};
    name = "user data function",
    dependencies::String = "",
    quadorder::Int = 0)

Provides a negotation interface for some data function with extended dependencies (region number, item number and local coordinates) that can be used in integrate!. The function f has to obey the interface

f(result, [X, T, R, I, L])

where the parameters X (= space coordinates) and T ( = time) are optional. Which of them are used has to be specified in the String dependencies. The string "X,I" for example specifies that the interface is only space-dependent and item-dependent, i.e

f(result, X, I)

The array dimensions specifies the expected length of result and X (if X-depdendent, otherwise will be ignored) and quadorder determines the additional quadrature order to be used if this function is involved in some quadrature-requireing procedure.

Action Kernels

Another for of user data are action kernel used to define an AbstractAction. Actions modify arguments of Assembly Patterns (usually all but the last one) and so allow parameter-depend assemblies. To use them, the user defines some kernel function for the action that has the interface

function action_kernel!(result,input,X,T,R,I,L)
    # result = modified input, e.g.
    # multiplication with some parameter that can depend on
    # X = space coordinates
    # T = time
    # R = region number
    # I = item number (cell, face or edge number depending on assembly type)
    # L = local coordinates on item reference domain
end

action_kernel! (generic function with 1 method)

During assembly, input (in general) takes the role of all non-testfunction arguments and the result vector will be the one that is multiplied with the testfunctions. Additionally, the kernel function can depend on X, T, R, I and L as specified above. Again, the dependencies of an action_kernel are stated via a string in the constructor that should be a substring of "XTRIL". Note, that the avoidance of the X-dependency spares the computation of the global coordinates of the quadrature points in the assembly loops. Moreover, time-dependency of an action kernel e.g. triggers reassembly of the associated PDEoperator in each time step.

GradientRobustMultiPhysics.ActionKernel — Function

function ActionKernel(
    f::Function,
    dimensions::Array{Int,1};
    name = "user action kernel",
    dependencies::String = "",
    quadorder::Int = 0)

Provides a negotation interface for some function that can be used in the Action constructor to define a user-defined operator action. The function has to obey the interface

f(result, input, [X, T, R, I, L])

where the parameters X (= space coordinates), T ( = time), R (= region number), I (= item number), L (= local coordinates) are optional. Which of them are used has to be specified in the String dependencies. The string "XT" for example specifies that the interface is only space and time dependent, i.e

f(result, input, X, T)

The input vector usually provides the FunctionOperator evaluations of (a subset of) the ansatz arguments of the assembly pattern where the action is used. The array dimensions specifies the expected length of result and input and quadorder determines the additional quadrature order to be used if this function (or its derived action) is involved in an assembly process.

GradientRobustMultiPhysics.NLActionKernel — Function

function NLActionKernel(
    f::Function,
    dimensions::Array{Int,1};
    name = "nonlinear user action kernel",
    dependencies::String = "",
    quadorder::Int = 0)

Provides a negotation interface for some function to be used as a nonlinear action kernel that can be used in the NonlinearOperator constructor without automatic differentiation. The function f has to obey the interface

f(result, input_current, input_ansatz)

and can be seen as a linearisation of a nonlinearity that can depend on the operator evaluation of the current iterate (inputcurrent) and, as usual, the operator evaluations of the ansatz function (inputansatz).

No further dependencies are allowed currently. Note, that this is a work-in-progress feature.

Actions

Actions are used by Assembly Patterns and consist of an action kernel plus some additional infratructure. To generate an action from an action kernel works via the following functions.

GradientRobustMultiPhysics.Action — Type

function Action(
    T::Type{<:Real},
    kernel::UserData{<:AbstractActionKernel};
    name::String = "user action"

Creates an Action from a given specified action kernel that then can be used in an assembly pattern. T specifies the number format that should match the number format of the used quadrature rules and grid coordinates in the mesh (usually Float64).

Moreover, there are some shortcut action constructors that can be used directly without defining an action kernel first.

GradientRobustMultiPhysics.MultiplyScalarAction — Function

function MultiplyScalarAction(value, ncomponents::Int)

Directly creates an Action that multiplies a scalar value to the input (vector of length ncomponents).

GradientRobustMultiPhysics.DoNotChangeAction — Function

function MultiplyScalarAction(value, ncomponents::Int)

Directly creates an Action that just copies the input to the result.