Export and Viewers

Plotting is possible e.g. via calculation of nodal values and the plot routines from ExtendableGrids.jl. It is also possible to export (the nodal values) of a FEVector as a vtk datafile and load it into e.g. paraview.

GradientRobustMultiPhysics.plot — Method

plot(Source::FEVector, blockids::Array{Int64,1}, operators::Array{DataType,1}; Plotter, use_subplots, subplots_per_column, colorlevels, isolines, aspect, show, clear, cbar, verbosity, cmap, maintitle, fsize)

Plots operators applied to components of a given FEVector via ExtendablePlots.plot, either in subplots or several single plots.

GradientRobustMultiPhysics.writeVTK! — Method

writeVTK!(filename::String, Data::FEVector; blocks, operators, names, vectorabs)

Writes the specified FEVector into a vtk datafile with the given filename. Each component of each FEVectorBlock (or the subset specified by blocks) is saved separately. Vector-valued quantities also generate a data field that represents the absolute value of the vector field at each grid point.